+++ frontend/src/Components/index.js
@@ -7,8 +7,6 @@ export { default as PostCardL } from "./PostCardL/PostCardL";
 export { default as PostCardS } from "./PostCardS/PostCardS";
 export { default as CreateCard } from "./CreateCard/CreateCard";
 export { default as QuestionCard } from "./QuestionCard/QuestionCard";
-export { default as QuestionCardGrid } from "./QuestionCardGrid/QuestionCardGrid";
 export { default as QuestionOpened } from "./QuestionOpened/QuestionOpened";
 export { default as TodayQuestion } from "./TodayQuestion/TodayQuestion";
-export { default as MusicCardGrid } from "./MusicCardGrid/MusicCardGrid";
 export { default as MusicCardOpened } from "./MusicCardOpened/MusicCardOpened";
