+++ frontend/static/frontend/main.js
@@ -437,28 +437,6 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 
 /***/ }),
 
-/***/ "./src/Components/MusicCardGrid/MusicCardGrid.elements.js":
-/*!****************************************************************!*\
-  !*** ./src/Components/MusicCardGrid/MusicCardGrid.elements.js ***!
-  \****************************************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"MusicListContainer\": () => (/* binding */ MusicListContainer),\n/* harmony export */   \"MusicCardWrapper\": () => (/* binding */ MusicCardWrapper)\n/* harmony export */ });\n/* harmony import */ var styled_components__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! styled-components */ \"./node_modules/styled-components/dist/styled-components.browser.esm.js\");\n/* harmony import */ var _styles_Colors__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../styles/Colors */ \"./src/styles/Colors.js\");\n/* harmony import */ var _styles_globalStyles__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../styles/globalStyles */ \"./src/styles/globalStyles.js\");\nvar _templateObject, _templateObject2;\n\nfunction _taggedTemplateLiteral(strings, raw) { if (!raw) { raw = strings.slice(0); } return Object.freeze(Object.defineProperties(strings, { raw: { value: Object.freeze(raw) } })); }\n\n\n\n\nvar MusicListContainer = styled_components__WEBPACK_IMPORTED_MODULE_2__[\"default\"].div(_templateObject || (_templateObject = _taggedTemplateLiteral([\"\\n  width: 100%;\\n  display: grid;\\n  grid-template-columns: repeat(3, 1fr);\\n  justify-items: center;\\n  gap: 5%;\\n  margin: 5% auto;\\n\"])));\nvar MusicCardWrapper = styled_components__WEBPACK_IMPORTED_MODULE_2__[\"default\"].div(_templateObject2 || (_templateObject2 = _taggedTemplateLiteral([\"\\n  cursor: pointer;\\n  width: 18rem;\\n  height: 7.5rem;\\n  display: flex;\\n  align-items: center;\\n  border: 1px solid \", \";\\n  border-radius: 5px;\\n  box-shadow: \", \";\\n  \", \"\\n\"])), _styles_Colors__WEBPACK_IMPORTED_MODULE_0__.cardStroke, _styles_globalStyles__WEBPACK_IMPORTED_MODULE_1__.CardShadow, function (p) {\n  return p.open === true ? \"border: 1px solid \".concat(_styles_Colors__WEBPACK_IMPORTED_MODULE_0__.iiPurple, \";\") : \"border: 1px solid \".concat(_styles_Colors__WEBPACK_IMPORTED_MODULE_0__.cardStroke, \";\");\n});\n\n//# sourceURL=webpack://frontend/./src/Components/MusicCardGrid/MusicCardGrid.elements.js?");
-
-/***/ }),
-
-/***/ "./src/Components/MusicCardGrid/MusicCardGrid.js":
-/*!*******************************************************!*\
-  !*** ./src/Components/MusicCardGrid/MusicCardGrid.js ***!
-  \*******************************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var _MusicCardGrid_elements__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./MusicCardGrid.elements */ \"./src/Components/MusicCardGrid/MusicCardGrid.elements.js\");\n/* harmony import */ var ___WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! .. */ \"./src/Components/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! axios */ \"./node_modules/axios/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_3___default = /*#__PURE__*/__webpack_require__.n(axios__WEBPACK_IMPORTED_MODULE_3__);\nfunction _slicedToArray(arr, i) { return _arrayWithHoles(arr) || _iterableToArrayLimit(arr, i) || _unsupportedIterableToArray(arr, i) || _nonIterableRest(); }\n\nfunction _nonIterableRest() { throw new TypeError(\"Invalid attempt to destructure non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); }\n\nfunction _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === \"string\") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === \"Object\" && o.constructor) n = o.constructor.name; if (n === \"Map\" || n === \"Set\") return Array.from(o); if (n === \"Arguments\" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }\n\nfunction _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }\n\nfunction _iterableToArrayLimit(arr, i) { var _i = arr == null ? null : typeof Symbol !== \"undefined\" && arr[Symbol.iterator] || arr[\"@@iterator\"]; if (_i == null) return; var _arr = []; var _n = true; var _d = false; var _s, _e; try { for (_i = _i.call(arr); !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i[\"return\"] != null) _i[\"return\"](); } finally { if (_d) throw _e; } } return _arr; }\n\nfunction _arrayWithHoles(arr) { if (Array.isArray(arr)) return arr; }\n\n\n\n\n\n\nvar MusicCardGrid = function MusicCardGrid(_ref) {\n  var currUser = _ref.currUser,\n      token = _ref.token,\n      userId = _ref.userId,\n      list = _ref.list;\n\n  var _useState = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(null),\n      _useState2 = _slicedToArray(_useState, 2),\n      content = _useState2[0],\n      setContent = _useState2[1];\n\n  var _useState3 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(0),\n      _useState4 = _slicedToArray(_useState3, 2),\n      openCard = _useState4[0],\n      setOpenCard = _useState4[1];\n\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    setContent();\n\n    if (list && list.length) {\n      setContent(list);\n    }\n  }, [list]);\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    if (!list) {\n      axios__WEBPACK_IMPORTED_MODULE_3___default()({\n        method: \"get\",\n        url: \"/api/recentmusic/6\",\n        headers: {\n          \"Content-Type\": \"application/json\"\n        }\n      }).then(function (response) {\n        if (response.status > 400) {\n          setPlaceholder(\"Something went wrong!\");\n        }\n\n        return response.data;\n      }).then(function (data) {\n        setContent(data[\"music list\"]);\n      });\n    }\n  }, []);\n\n  var handleClick = function handleClick(clickedCard) {\n    setOpenCard(clickedCard);\n  };\n\n  var MusicCardList = content ? content.map(function (c, i) {\n    return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_MusicCardGrid_elements__WEBPACK_IMPORTED_MODULE_1__.MusicCardWrapper, {\n      key: i,\n      onclick: function onclick() {\n        return handleClick(i);\n      },\n      open: openCard === i\n    }, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(___WEBPACK_IMPORTED_MODULE_2__.MusicCard, {\n      title: c[0],\n      artist: c[1],\n      cover: c[2],\n      source: c[3],\n      link: c[4],\n      cardIndex: i,\n      handleClick: handleClick,\n      open: openCard === i\n    }));\n  }) : [];\n  return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(react__WEBPACK_IMPORTED_MODULE_0__.Fragment, null, content ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(react__WEBPACK_IMPORTED_MODULE_0__.Fragment, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_MusicCardGrid_elements__WEBPACK_IMPORTED_MODULE_1__.MusicListContainer, null, MusicCardList), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(___WEBPACK_IMPORTED_MODULE_2__.MusicCardOpened, {\n    currUser: currUser,\n    token: token,\n    userId: userId,\n    cardIndex: openCard,\n    title: content[openCard] ? content[openCard][0] : \"\",\n    artist: content[openCard] ? content[openCard][1] : \"\",\n    cover: content[openCard] ? content[openCard][2] : \"\",\n    source: content[openCard] ? content[openCard][3] : \"\",\n    link: content[openCard] ? content[openCard][4] : \"\"\n  })) : \"\");\n};\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (MusicCardGrid);\n\n//# sourceURL=webpack://frontend/./src/Components/MusicCardGrid/MusicCardGrid.js?");
-
-/***/ }),
-
 /***/ "./src/Components/MusicCardOpened/MusicCardOpened.elements.js":
 /*!********************************************************************!*\
   !*** ./src/Components/MusicCardOpened/MusicCardOpened.elements.js ***!
@@ -566,28 +544,6 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 
 "use strict";
 eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var _PostCardS_elements__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./PostCardS.elements */ \"./src/Components/PostCardS/PostCardS.elements.js\");\n\n\n\nvar PostCardS = function PostCardS(_ref) {\n  var id = _ref.id,\n      user = _ref.user,\n      title = _ref.title,\n      track_title = _ref.track_title,\n      track_artist = _ref.track_artist,\n      track_album_cover = _ref.track_album_cover,\n      handleCardOpen = _ref.handleCardOpen;\n  return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_PostCardS_elements__WEBPACK_IMPORTED_MODULE_1__.PostCardArea, {\n    onClick: function onClick() {\n      return handleCardOpen(id);\n    }\n  }, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_PostCardS_elements__WEBPACK_IMPORTED_MODULE_1__.TopArea, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_PostCardS_elements__WEBPACK_IMPORTED_MODULE_1__.Author, null, user)), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_PostCardS_elements__WEBPACK_IMPORTED_MODULE_1__.MusicArea, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_PostCardS_elements__WEBPACK_IMPORTED_MODULE_1__.MusicCover, {\n    src: track_album_cover\n  }), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_PostCardS_elements__WEBPACK_IMPORTED_MODULE_1__.MusicInfoArea, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_PostCardS_elements__WEBPACK_IMPORTED_MODULE_1__.MusicTitleWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_PostCardS_elements__WEBPACK_IMPORTED_MODULE_1__.MusicTitle, {\n    slide: track_title.length > 10\n  }, track_title)), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_PostCardS_elements__WEBPACK_IMPORTED_MODULE_1__.MusicArtistWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_PostCardS_elements__WEBPACK_IMPORTED_MODULE_1__.MusicArtist, {\n    slide: track_artist.length > 10\n  }, track_artist)))), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_PostCardS_elements__WEBPACK_IMPORTED_MODULE_1__.TitleArea, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_PostCardS_elements__WEBPACK_IMPORTED_MODULE_1__.ContentTitle, null, title.length > 30 ? title.slice(0, 30) + \"...\" : title)));\n};\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (PostCardS);\n\n//# sourceURL=webpack://frontend/./src/Components/PostCardS/PostCardS.js?");
-
-/***/ }),
-
-/***/ "./src/Components/QuestionCardGrid/QuestionCardGrid.elements.js":
-/*!**********************************************************************!*\
-  !*** ./src/Components/QuestionCardGrid/QuestionCardGrid.elements.js ***!
-  \**********************************************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"QuestionGridContainer\": () => (/* binding */ QuestionGridContainer),\n/* harmony export */   \"QuestionCardWrapper\": () => (/* binding */ QuestionCardWrapper)\n/* harmony export */ });\n/* harmony import */ var styled_components__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! styled-components */ \"./node_modules/styled-components/dist/styled-components.browser.esm.js\");\n/* harmony import */ var _styles_Colors__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../styles/Colors */ \"./src/styles/Colors.js\");\n/* harmony import */ var _styles_globalStyles__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../styles/globalStyles */ \"./src/styles/globalStyles.js\");\nvar _templateObject, _templateObject2;\n\nfunction _taggedTemplateLiteral(strings, raw) { if (!raw) { raw = strings.slice(0); } return Object.freeze(Object.defineProperties(strings, { raw: { value: Object.freeze(raw) } })); }\n\n\n\n\nvar QuestionGridContainer = styled_components__WEBPACK_IMPORTED_MODULE_2__[\"default\"].div(_templateObject || (_templateObject = _taggedTemplateLiteral([\"\\n  display: grid;\\n  grid-template-columns: repeat(3, 29vw);\\n  grid-gap: 1.5vw;\\n  margin: 3% auto;\\n  width: 100%;\\n\\n  @media screen and (min-width: 1920px) {\\n    grid-template-columns: repeat(3, 19vw);\\n  }\\n\"])));\nvar QuestionCardWrapper = styled_components__WEBPACK_IMPORTED_MODULE_2__[\"default\"].div(_templateObject2 || (_templateObject2 = _taggedTemplateLiteral([\"\\n  text-align: center;\\n\"])));\n\n//# sourceURL=webpack://frontend/./src/Components/QuestionCardGrid/QuestionCardGrid.elements.js?");
-
-/***/ }),
-
-/***/ "./src/Components/QuestionCardGrid/QuestionCardGrid.js":
-/*!*************************************************************!*\
-  !*** ./src/Components/QuestionCardGrid/QuestionCardGrid.js ***!
-  \*************************************************************/
-/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
-
-"use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var _QuestionCardGrid_elements__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./QuestionCardGrid.elements */ \"./src/Components/QuestionCardGrid/QuestionCardGrid.elements.js\");\n/* harmony import */ var ___WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! .. */ \"./src/Components/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! axios */ \"./node_modules/axios/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_3___default = /*#__PURE__*/__webpack_require__.n(axios__WEBPACK_IMPORTED_MODULE_3__);\nfunction _slicedToArray(arr, i) { return _arrayWithHoles(arr) || _iterableToArrayLimit(arr, i) || _unsupportedIterableToArray(arr, i) || _nonIterableRest(); }\n\nfunction _nonIterableRest() { throw new TypeError(\"Invalid attempt to destructure non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); }\n\nfunction _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === \"string\") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === \"Object\" && o.constructor) n = o.constructor.name; if (n === \"Map\" || n === \"Set\") return Array.from(o); if (n === \"Arguments\" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }\n\nfunction _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }\n\nfunction _iterableToArrayLimit(arr, i) { var _i = arr == null ? null : typeof Symbol !== \"undefined\" && arr[Symbol.iterator] || arr[\"@@iterator\"]; if (_i == null) return; var _arr = []; var _n = true; var _d = false; var _s, _e; try { for (_i = _i.call(arr); !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i[\"return\"] != null) _i[\"return\"](); } finally { if (_d) throw _e; } } return _arr; }\n\nfunction _arrayWithHoles(arr) { if (Array.isArray(arr)) return arr; }\n\n\n\n\n\n\nvar QuestionCardGrid = function QuestionCardGrid(_ref) {\n  var currUser = _ref.currUser,\n      token = _ref.token,\n      userId = _ref.userId,\n      list = _ref.list;\n\n  var _useState = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(),\n      _useState2 = _slicedToArray(_useState, 2),\n      content = _useState2[0],\n      setContent = _useState2[1];\n\n  var _useState3 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(-1),\n      _useState4 = _slicedToArray(_useState3, 2),\n      openCard = _useState4[0],\n      setOpenCard = _useState4[1];\n\n  var _useState5 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(false),\n      _useState6 = _slicedToArray(_useState5, 2),\n      isOpened = _useState6[0],\n      setIsOpened = _useState6[1];\n\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    setContent();\n\n    if (list && list.length) {\n      setContent(list);\n    }\n  }, [list]);\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    if (!list) {\n      axios__WEBPACK_IMPORTED_MODULE_3___default()({\n        method: \"get\",\n        url: \"/api/question\",\n        headers: {\n          \"Content-Type\": \"application/json\"\n        }\n      }).then(function (response) {\n        if (response.status > 400) {\n          setPlaceholder(\"Something went wrong!\");\n        }\n\n        return response.data;\n      }).then(function (data) {\n        setContent(data);\n      });\n    }\n  }, []);\n\n  var handleClick = function handleClick(clickedCard) {\n    if (clickedCard === openCard) {\n      setOpenCard(-1);\n      setIsOpened(false);\n    } else {\n      //다른 카드 클릭\n      setOpenCard(clickedCard);\n      setIsOpened(true);\n    }\n  };\n\n  var QuestionCardList = content ? content.map(function (c, i) {\n    return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_QuestionCardGrid_elements__WEBPACK_IMPORTED_MODULE_1__.QuestionCardWrapper, {\n      key: i\n    }, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(___WEBPACK_IMPORTED_MODULE_2__.QuestionCard, {\n      question: c.question_content,\n      date: c.released_date,\n      cardIndex: i,\n      handleClick: handleClick,\n      open: isOpened && openCard === i\n    }), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(___WEBPACK_IMPORTED_MODULE_2__.QuestionOpened, {\n      currUser: currUser,\n      token: token,\n      userId: userId,\n      questionId: c.id,\n      question: c.question_content,\n      detail: c.explain,\n      cardIndex: i,\n      open: isOpened && openCard === i\n    }));\n  }) : [];\n  return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_QuestionCardGrid_elements__WEBPACK_IMPORTED_MODULE_1__.QuestionGridContainer, null, QuestionCardList);\n};\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (QuestionCardGrid);\n\n//# sourceURL=webpack://frontend/./src/Components/QuestionCardGrid/QuestionCardGrid.js?");
 
 /***/ }),
 
@@ -708,7 +664,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"MusicPlayer\": () => (/* reexport safe */ _MusicPlayer_MusicPlayer__WEBPACK_IMPORTED_MODULE_0__[\"default\"]),\n/* harmony export */   \"MusicCard\": () => (/* reexport safe */ _MusicCard_MusicCard__WEBPACK_IMPORTED_MODULE_1__[\"default\"]),\n/* harmony export */   \"Calendar\": () => (/* reexport safe */ _Calendar_Calendar__WEBPACK_IMPORTED_MODULE_2__[\"default\"]),\n/* harmony export */   \"List\": () => (/* reexport safe */ _List_List__WEBPACK_IMPORTED_MODULE_3__[\"default\"]),\n/* harmony export */   \"SearchedItem\": () => (/* reexport safe */ _SearchedItem_SearchedItem__WEBPACK_IMPORTED_MODULE_4__[\"default\"]),\n/* harmony export */   \"PostCardL\": () => (/* reexport safe */ _PostCardL_PostCardL__WEBPACK_IMPORTED_MODULE_5__[\"default\"]),\n/* harmony export */   \"PostCardS\": () => (/* reexport safe */ _PostCardS_PostCardS__WEBPACK_IMPORTED_MODULE_6__[\"default\"]),\n/* harmony export */   \"CreateCard\": () => (/* reexport safe */ _CreateCard_CreateCard__WEBPACK_IMPORTED_MODULE_7__[\"default\"]),\n/* harmony export */   \"QuestionCard\": () => (/* reexport safe */ _QuestionCard_QuestionCard__WEBPACK_IMPORTED_MODULE_8__[\"default\"]),\n/* harmony export */   \"QuestionCardGrid\": () => (/* reexport safe */ _QuestionCardGrid_QuestionCardGrid__WEBPACK_IMPORTED_MODULE_9__[\"default\"]),\n/* harmony export */   \"QuestionOpened\": () => (/* reexport safe */ _QuestionOpened_QuestionOpened__WEBPACK_IMPORTED_MODULE_10__[\"default\"]),\n/* harmony export */   \"TodayQuestion\": () => (/* reexport safe */ _TodayQuestion_TodayQuestion__WEBPACK_IMPORTED_MODULE_11__[\"default\"]),\n/* harmony export */   \"MusicCardGrid\": () => (/* reexport safe */ _MusicCardGrid_MusicCardGrid__WEBPACK_IMPORTED_MODULE_12__[\"default\"]),\n/* harmony export */   \"MusicCardOpened\": () => (/* reexport safe */ _MusicCardOpened_MusicCardOpened__WEBPACK_IMPORTED_MODULE_13__[\"default\"])\n/* harmony export */ });\n/* harmony import */ var _MusicPlayer_MusicPlayer__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./MusicPlayer/MusicPlayer */ \"./src/Components/MusicPlayer/MusicPlayer.js\");\n/* harmony import */ var _MusicCard_MusicCard__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./MusicCard/MusicCard */ \"./src/Components/MusicCard/MusicCard.js\");\n/* harmony import */ var _Calendar_Calendar__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./Calendar/Calendar */ \"./src/Components/Calendar/Calendar.js\");\n/* harmony import */ var _List_List__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./List/List */ \"./src/Components/List/List.js\");\n/* harmony import */ var _SearchedItem_SearchedItem__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./SearchedItem/SearchedItem */ \"./src/Components/SearchedItem/SearchedItem.js\");\n/* harmony import */ var _PostCardL_PostCardL__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ./PostCardL/PostCardL */ \"./src/Components/PostCardL/PostCardL.js\");\n/* harmony import */ var _PostCardS_PostCardS__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ./PostCardS/PostCardS */ \"./src/Components/PostCardS/PostCardS.js\");\n/* harmony import */ var _CreateCard_CreateCard__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! ./CreateCard/CreateCard */ \"./src/Components/CreateCard/CreateCard.js\");\n/* harmony import */ var _QuestionCard_QuestionCard__WEBPACK_IMPORTED_MODULE_8__ = __webpack_require__(/*! ./QuestionCard/QuestionCard */ \"./src/Components/QuestionCard/QuestionCard.js\");\n/* harmony import */ var _QuestionCardGrid_QuestionCardGrid__WEBPACK_IMPORTED_MODULE_9__ = __webpack_require__(/*! ./QuestionCardGrid/QuestionCardGrid */ \"./src/Components/QuestionCardGrid/QuestionCardGrid.js\");\n/* harmony import */ var _QuestionOpened_QuestionOpened__WEBPACK_IMPORTED_MODULE_10__ = __webpack_require__(/*! ./QuestionOpened/QuestionOpened */ \"./src/Components/QuestionOpened/QuestionOpened.js\");\n/* harmony import */ var _TodayQuestion_TodayQuestion__WEBPACK_IMPORTED_MODULE_11__ = __webpack_require__(/*! ./TodayQuestion/TodayQuestion */ \"./src/Components/TodayQuestion/TodayQuestion.js\");\n/* harmony import */ var _MusicCardGrid_MusicCardGrid__WEBPACK_IMPORTED_MODULE_12__ = __webpack_require__(/*! ./MusicCardGrid/MusicCardGrid */ \"./src/Components/MusicCardGrid/MusicCardGrid.js\");\n/* harmony import */ var _MusicCardOpened_MusicCardOpened__WEBPACK_IMPORTED_MODULE_13__ = __webpack_require__(/*! ./MusicCardOpened/MusicCardOpened */ \"./src/Components/MusicCardOpened/MusicCardOpened.js\");\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n//# sourceURL=webpack://frontend/./src/Components/index.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"MusicPlayer\": () => (/* reexport safe */ _MusicPlayer_MusicPlayer__WEBPACK_IMPORTED_MODULE_0__[\"default\"]),\n/* harmony export */   \"MusicCard\": () => (/* reexport safe */ _MusicCard_MusicCard__WEBPACK_IMPORTED_MODULE_1__[\"default\"]),\n/* harmony export */   \"Calendar\": () => (/* reexport safe */ _Calendar_Calendar__WEBPACK_IMPORTED_MODULE_2__[\"default\"]),\n/* harmony export */   \"List\": () => (/* reexport safe */ _List_List__WEBPACK_IMPORTED_MODULE_3__[\"default\"]),\n/* harmony export */   \"SearchedItem\": () => (/* reexport safe */ _SearchedItem_SearchedItem__WEBPACK_IMPORTED_MODULE_4__[\"default\"]),\n/* harmony export */   \"PostCardL\": () => (/* reexport safe */ _PostCardL_PostCardL__WEBPACK_IMPORTED_MODULE_5__[\"default\"]),\n/* harmony export */   \"PostCardS\": () => (/* reexport safe */ _PostCardS_PostCardS__WEBPACK_IMPORTED_MODULE_6__[\"default\"]),\n/* harmony export */   \"CreateCard\": () => (/* reexport safe */ _CreateCard_CreateCard__WEBPACK_IMPORTED_MODULE_7__[\"default\"]),\n/* harmony export */   \"QuestionCard\": () => (/* reexport safe */ _QuestionCard_QuestionCard__WEBPACK_IMPORTED_MODULE_8__[\"default\"]),\n/* harmony export */   \"QuestionOpened\": () => (/* reexport safe */ _QuestionOpened_QuestionOpened__WEBPACK_IMPORTED_MODULE_9__[\"default\"]),\n/* harmony export */   \"TodayQuestion\": () => (/* reexport safe */ _TodayQuestion_TodayQuestion__WEBPACK_IMPORTED_MODULE_10__[\"default\"]),\n/* harmony export */   \"MusicCardOpened\": () => (/* reexport safe */ _MusicCardOpened_MusicCardOpened__WEBPACK_IMPORTED_MODULE_11__[\"default\"])\n/* harmony export */ });\n/* harmony import */ var _MusicPlayer_MusicPlayer__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./MusicPlayer/MusicPlayer */ \"./src/Components/MusicPlayer/MusicPlayer.js\");\n/* harmony import */ var _MusicCard_MusicCard__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./MusicCard/MusicCard */ \"./src/Components/MusicCard/MusicCard.js\");\n/* harmony import */ var _Calendar_Calendar__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./Calendar/Calendar */ \"./src/Components/Calendar/Calendar.js\");\n/* harmony import */ var _List_List__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./List/List */ \"./src/Components/List/List.js\");\n/* harmony import */ var _SearchedItem_SearchedItem__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./SearchedItem/SearchedItem */ \"./src/Components/SearchedItem/SearchedItem.js\");\n/* harmony import */ var _PostCardL_PostCardL__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ./PostCardL/PostCardL */ \"./src/Components/PostCardL/PostCardL.js\");\n/* harmony import */ var _PostCardS_PostCardS__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ./PostCardS/PostCardS */ \"./src/Components/PostCardS/PostCardS.js\");\n/* harmony import */ var _CreateCard_CreateCard__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! ./CreateCard/CreateCard */ \"./src/Components/CreateCard/CreateCard.js\");\n/* harmony import */ var _QuestionCard_QuestionCard__WEBPACK_IMPORTED_MODULE_8__ = __webpack_require__(/*! ./QuestionCard/QuestionCard */ \"./src/Components/QuestionCard/QuestionCard.js\");\n/* harmony import */ var _QuestionOpened_QuestionOpened__WEBPACK_IMPORTED_MODULE_9__ = __webpack_require__(/*! ./QuestionOpened/QuestionOpened */ \"./src/Components/QuestionOpened/QuestionOpened.js\");\n/* harmony import */ var _TodayQuestion_TodayQuestion__WEBPACK_IMPORTED_MODULE_10__ = __webpack_require__(/*! ./TodayQuestion/TodayQuestion */ \"./src/Components/TodayQuestion/TodayQuestion.js\");\n/* harmony import */ var _MusicCardOpened_MusicCardOpened__WEBPACK_IMPORTED_MODULE_11__ = __webpack_require__(/*! ./MusicCardOpened/MusicCardOpened */ \"./src/Components/MusicCardOpened/MusicCardOpened.js\");\n\n\n\n\n\n\n\n\n\n\n\n\n\n//# sourceURL=webpack://frontend/./src/Components/index.js?");
 
 /***/ }),
 
@@ -774,7 +730,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var _styles_globalStyles__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../styles/globalStyles */ \"./src/styles/globalStyles.js\");\n/* harmony import */ var _Components__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../Components */ \"./src/Components/index.js\");\n/* harmony import */ var _Explore_elements__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./Explore.elements */ \"./src/Containers/Explore/Explore.elements.js\");\n\n\n\n\n\n\nvar Explore = function Explore(_ref) {\n  var currUser = _ref.currUser,\n      token = _ref.token,\n      userId = _ref.userId;\n  return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(react__WEBPACK_IMPORTED_MODULE_0__.Fragment, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_styles_globalStyles__WEBPACK_IMPORTED_MODULE_1__.Background, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_styles_globalStyles__WEBPACK_IMPORTED_MODULE_1__.PageSection, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_styles_globalStyles__WEBPACK_IMPORTED_MODULE_1__.PageTitleWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_styles_globalStyles__WEBPACK_IMPORTED_MODULE_1__.PageTitle, null, \"\\uC9C8\\uBB38 \\uB458\\uB7EC\\uBCF4\\uAE30\")), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Explore_elements__WEBPACK_IMPORTED_MODULE_3__.QuestionListWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_2__.QuestionCardGrid, {\n    currUser: currUser,\n    token: token,\n    userId: userId\n  })))));\n};\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (Explore);\n\n//# sourceURL=webpack://frontend/./src/Containers/Explore/Explore.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var _styles_globalStyles__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../styles/globalStyles */ \"./src/styles/globalStyles.js\");\n/* harmony import */ var ___WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../ */ \"./src/Containers/index.js\");\n/* harmony import */ var _Explore_elements__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./Explore.elements */ \"./src/Containers/Explore/Explore.elements.js\");\n\n\n\n\n\nvar Explore = function Explore(_ref) {\n  var currUser = _ref.currUser,\n      token = _ref.token,\n      userId = _ref.userId;\n  return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(react__WEBPACK_IMPORTED_MODULE_0__.Fragment, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_styles_globalStyles__WEBPACK_IMPORTED_MODULE_1__.Background, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_styles_globalStyles__WEBPACK_IMPORTED_MODULE_1__.PageSection, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_styles_globalStyles__WEBPACK_IMPORTED_MODULE_1__.PageTitleWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_styles_globalStyles__WEBPACK_IMPORTED_MODULE_1__.PageTitle, null, \"\\uC9C8\\uBB38 \\uB458\\uB7EC\\uBCF4\\uAE30\")), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Explore_elements__WEBPACK_IMPORTED_MODULE_3__.QuestionListWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(___WEBPACK_IMPORTED_MODULE_2__.QuestionCardGrid, {\n    currUser: currUser,\n    token: token,\n    userId: userId\n  })))));\n};\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (Explore);\n\n//# sourceURL=webpack://frontend/./src/Containers/Explore/Explore.js?");
 
 /***/ }),
 
@@ -818,7 +774,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! axios */ \"./node_modules/axios/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(axios__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var _Home_elements__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./Home.elements */ \"./src/Containers/Home/Home.elements.js\");\n/* harmony import */ var _styles_globalStyles__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ../../styles/globalStyles */ \"./src/styles/globalStyles.js\");\n/* harmony import */ var _Components__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ../../Components */ \"./src/Components/index.js\");\nfunction _slicedToArray(arr, i) { return _arrayWithHoles(arr) || _iterableToArrayLimit(arr, i) || _unsupportedIterableToArray(arr, i) || _nonIterableRest(); }\n\nfunction _nonIterableRest() { throw new TypeError(\"Invalid attempt to destructure non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); }\n\nfunction _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === \"string\") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === \"Object\" && o.constructor) n = o.constructor.name; if (n === \"Map\" || n === \"Set\") return Array.from(o); if (n === \"Arguments\" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }\n\nfunction _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }\n\nfunction _iterableToArrayLimit(arr, i) { var _i = arr == null ? null : typeof Symbol !== \"undefined\" && arr[Symbol.iterator] || arr[\"@@iterator\"]; if (_i == null) return; var _arr = []; var _n = true; var _d = false; var _s, _e; try { for (_i = _i.call(arr); !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i[\"return\"] != null) _i[\"return\"](); } finally { if (_d) throw _e; } } return _arr; }\n\nfunction _arrayWithHoles(arr) { if (Array.isArray(arr)) return arr; }\n\n\n\n\n\n\n\nvar Home = function Home(_ref) {\n  var currUser = _ref.currUser,\n      token = _ref.token,\n      userId = _ref.userId;\n\n  var _useState = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(false),\n      _useState2 = _slicedToArray(_useState, 2),\n      loaded = _useState2[0],\n      setLoad = _useState2[1];\n\n  var _useState3 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(\"Loading Content\"),\n      _useState4 = _slicedToArray(_useState3, 2),\n      placeholder = _useState4[0],\n      setPlaceholder = _useState4[1];\n\n  var _useState5 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(),\n      _useState6 = _slicedToArray(_useState5, 2),\n      pageQuestion = _useState6[0],\n      setPageQuestion = _useState6[1];\n\n  return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(react__WEBPACK_IMPORTED_MODULE_0__.Fragment, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_styles_globalStyles__WEBPACK_IMPORTED_MODULE_3__.Background, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.TodayQuestionSection, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_4__.TodayQuestion, {\n    currUser: currUser,\n    token: token,\n    userId: userId,\n    setPageQuestion: setPageQuestion\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.CreatePageSection, {\n    id: \"create\"\n  }, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.CreateCardLeft, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.HelperLeft, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uB2F9\\uC2E0\\uC758 \\uC774\\uC57C\\uAE30\\uB97C \\uB4E4\\uB824\\uC8FC\\uC138\\uC694.\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uC5B4\\uB5A4 \\uB0B4\\uC6A9\\uC774\\uB4E0 \\uC88B\\uC544\\uC694. \", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uC0AC\\uC18C\\uD55C \\uC77C\\uC0C1\\uBD80\\uD130 \\uAE4A\\uC740 \\uC18D\\uB9C8\\uC74C\\uAE4C\\uC9C0,\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uB5A0\\uC624\\uB974\\uB294 \\uB300\\uB85C \\uC801\\uC5B4\\uBCFC\\uAE4C\\uC694?\")), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_4__.CreateCard, {\n    currUser: currUser,\n    token: token,\n    userId: userId,\n    questionId: pageQuestion,\n    locationAt: \"home\"\n  }), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.CreateCardRight, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.HelperQuestionArea, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.HelperRight, null, \"\\uC9C8\\uBB38\\uC744 \\uB4E3\\uACE0,\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uB5A0\\uC624\\uB974\\uB294 \\uC74C\\uC545\\uC774\\uB098 \\uC774\\uC57C\\uAE30\\uAC00 \\uC788\\uB098\\uC694?\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uB2F9\\uC2E0\\uC758 \\uC774\\uC57C\\uAE30\\uC5D0 \\uC5B4\\uC6B8\\uB9AC\\uB294 \\uC74C\\uC545\\uC744\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uC9C1\\uC811 \\uACE8\\uB77C\\uC8FC\\uC138\\uC694.\")), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.DoneButtonArea, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.HelperRight, null, \"\\uC774\\uC81C \\uB2F9\\uC2E0\\uC758 \\uC774\\uC57C\\uAE30\\uB97C\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uB2E4\\uB978 \\uC0AC\\uB78C\\uB4E4\\uACFC \\uB098\\uB220\\uBCF4\\uC138\\uC694!\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null))))), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.QuestionListPageSection, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.PageHeaderContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.PageHeader, null, \"\\uC774\\uB7F0 \\uC9C8\\uBB38\\uC740 \\uC5B4\\uB54C\\uC694?\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.ToQuestionListButton, {\n    to: \"/explore\"\n  }, \"\\uBAA8\\uB4E0 \\uC9C8\\uBB38 \\uBCF4\\uAE30 >\")), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_4__.QuestionCardGrid, {\n    currUser: currUser,\n    token: token,\n    userId: userId\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.MusicListPageSection, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.PageHeaderContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.PageHeader, null, \"\\uCD5C\\uADFC\\uC5D0 \\uC120\\uD0DD\\uB41C \\uC74C\\uC545\\uC774\\uC5D0\\uC694.\")), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_4__.MusicCardGrid, {\n    currUser: currUser,\n    token: token,\n    userId: userId\n  }))));\n};\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (Home);\n\n//# sourceURL=webpack://frontend/./src/Containers/Home/Home.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! axios */ \"./node_modules/axios/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(axios__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var _Home_elements__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./Home.elements */ \"./src/Containers/Home/Home.elements.js\");\n/* harmony import */ var _styles_globalStyles__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ../../styles/globalStyles */ \"./src/styles/globalStyles.js\");\n/* harmony import */ var _Components__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ../../Components */ \"./src/Components/index.js\");\n/* harmony import */ var ___WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ../ */ \"./src/Containers/index.js\");\nfunction _slicedToArray(arr, i) { return _arrayWithHoles(arr) || _iterableToArrayLimit(arr, i) || _unsupportedIterableToArray(arr, i) || _nonIterableRest(); }\n\nfunction _nonIterableRest() { throw new TypeError(\"Invalid attempt to destructure non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); }\n\nfunction _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === \"string\") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === \"Object\" && o.constructor) n = o.constructor.name; if (n === \"Map\" || n === \"Set\") return Array.from(o); if (n === \"Arguments\" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }\n\nfunction _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }\n\nfunction _iterableToArrayLimit(arr, i) { var _i = arr == null ? null : typeof Symbol !== \"undefined\" && arr[Symbol.iterator] || arr[\"@@iterator\"]; if (_i == null) return; var _arr = []; var _n = true; var _d = false; var _s, _e; try { for (_i = _i.call(arr); !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i[\"return\"] != null) _i[\"return\"](); } finally { if (_d) throw _e; } } return _arr; }\n\nfunction _arrayWithHoles(arr) { if (Array.isArray(arr)) return arr; }\n\n\n\n\n\n\n\n\nvar Home = function Home(_ref) {\n  var currUser = _ref.currUser,\n      token = _ref.token,\n      userId = _ref.userId;\n\n  var _useState = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(false),\n      _useState2 = _slicedToArray(_useState, 2),\n      loaded = _useState2[0],\n      setLoad = _useState2[1];\n\n  var _useState3 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(\"Loading Content\"),\n      _useState4 = _slicedToArray(_useState3, 2),\n      placeholder = _useState4[0],\n      setPlaceholder = _useState4[1];\n\n  var _useState5 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(),\n      _useState6 = _slicedToArray(_useState5, 2),\n      pageQuestion = _useState6[0],\n      setPageQuestion = _useState6[1];\n\n  return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(react__WEBPACK_IMPORTED_MODULE_0__.Fragment, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_styles_globalStyles__WEBPACK_IMPORTED_MODULE_3__.Background, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.TodayQuestionSection, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_4__.TodayQuestion, {\n    currUser: currUser,\n    token: token,\n    userId: userId,\n    setPageQuestion: setPageQuestion\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.CreatePageSection, {\n    id: \"create\"\n  }, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.CreateCardLeft, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.HelperLeft, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uB2F9\\uC2E0\\uC758 \\uC774\\uC57C\\uAE30\\uB97C \\uB4E4\\uB824\\uC8FC\\uC138\\uC694.\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uC5B4\\uB5A4 \\uB0B4\\uC6A9\\uC774\\uB4E0 \\uC88B\\uC544\\uC694. \", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uC0AC\\uC18C\\uD55C \\uC77C\\uC0C1\\uBD80\\uD130 \\uAE4A\\uC740 \\uC18D\\uB9C8\\uC74C\\uAE4C\\uC9C0,\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uB5A0\\uC624\\uB974\\uB294 \\uB300\\uB85C \\uC801\\uC5B4\\uBCFC\\uAE4C\\uC694?\")), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_4__.CreateCard, {\n    currUser: currUser,\n    token: token,\n    userId: userId,\n    questionId: pageQuestion,\n    locationAt: \"home\"\n  }), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.CreateCardRight, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.HelperQuestionArea, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.HelperRight, null, \"\\uC9C8\\uBB38\\uC744 \\uB4E3\\uACE0,\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uB5A0\\uC624\\uB974\\uB294 \\uC74C\\uC545\\uC774\\uB098 \\uC774\\uC57C\\uAE30\\uAC00 \\uC788\\uB098\\uC694?\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uB2F9\\uC2E0\\uC758 \\uC774\\uC57C\\uAE30\\uC5D0 \\uC5B4\\uC6B8\\uB9AC\\uB294 \\uC74C\\uC545\\uC744\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uC9C1\\uC811 \\uACE8\\uB77C\\uC8FC\\uC138\\uC694.\")), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.DoneButtonArea, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.HelperRight, null, \"\\uC774\\uC81C \\uB2F9\\uC2E0\\uC758 \\uC774\\uC57C\\uAE30\\uB97C\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uB2E4\\uB978 \\uC0AC\\uB78C\\uB4E4\\uACFC \\uB098\\uB220\\uBCF4\\uC138\\uC694!\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null))))), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.QuestionListPageSection, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.PageHeaderContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.PageHeader, null, \"\\uC774\\uB7F0 \\uC9C8\\uBB38\\uC740 \\uC5B4\\uB54C\\uC694?\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.ToQuestionListButton, {\n    to: \"/explore\"\n  }, \"\\uBAA8\\uB4E0 \\uC9C8\\uBB38 \\uBCF4\\uAE30 >\")), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(___WEBPACK_IMPORTED_MODULE_5__.QuestionCardGrid, {\n    currUser: currUser,\n    token: token,\n    userId: userId\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.MusicListPageSection, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.PageHeaderContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.PageHeader, null, \"\\uCD5C\\uADFC\\uC5D0 \\uC120\\uD0DD\\uB41C \\uC74C\\uC545\\uC774\\uC5D0\\uC694.\")), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(___WEBPACK_IMPORTED_MODULE_5__.MusicCardGrid, {\n    currUser: currUser,\n    token: token,\n    userId: userId\n  }))));\n};\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (Home);\n\n//# sourceURL=webpack://frontend/./src/Containers/Home/Home.js?");
 
 /***/ }),
 
@@ -841,6 +797,28 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 
 "use strict";
 eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! axios */ \"./node_modules/axios/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(axios__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var _Components_csrftoken__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../Components/csrftoken */ \"./src/Components/csrftoken.js\");\n/* harmony import */ var react_router__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! react-router */ \"./node_modules/react-router/esm/react-router.js\");\n/* harmony import */ var _styles_globalStyles__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ../../styles/globalStyles */ \"./src/styles/globalStyles.js\");\n/* harmony import */ var _Login_elements__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./Login.elements */ \"./src/Containers/Login/Login.elements.js\");\n/* harmony import */ var _SignUp_SignUp_elements__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ../SignUp/SignUp.elements */ \"./src/Containers/SignUp/SignUp.elements.js\");\nfunction _slicedToArray(arr, i) { return _arrayWithHoles(arr) || _iterableToArrayLimit(arr, i) || _unsupportedIterableToArray(arr, i) || _nonIterableRest(); }\n\nfunction _nonIterableRest() { throw new TypeError(\"Invalid attempt to destructure non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); }\n\nfunction _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === \"string\") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === \"Object\" && o.constructor) n = o.constructor.name; if (n === \"Map\" || n === \"Set\") return Array.from(o); if (n === \"Arguments\" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }\n\nfunction _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }\n\nfunction _iterableToArrayLimit(arr, i) { var _i = arr == null ? null : typeof Symbol !== \"undefined\" && arr[Symbol.iterator] || arr[\"@@iterator\"]; if (_i == null) return; var _arr = []; var _n = true; var _d = false; var _s, _e; try { for (_i = _i.call(arr); !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i[\"return\"] != null) _i[\"return\"](); } finally { if (_d) throw _e; } } return _arr; }\n\nfunction _arrayWithHoles(arr) { if (Array.isArray(arr)) return arr; }\n\n\n\n\n\n\n\n\n\nvar Login = function Login(_ref) {\n  var saveUserData = _ref.saveUserData,\n      currUser = _ref.currUser,\n      setNickname = _ref.setNickname;\n\n  var _useState = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(\"\"),\n      _useState2 = _slicedToArray(_useState, 2),\n      username = _useState2[0],\n      setUsername = _useState2[1];\n\n  var _useState3 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(\"\"),\n      _useState4 = _slicedToArray(_useState3, 2),\n      pwd = _useState4[0],\n      setPwd = _useState4[1];\n\n  var _useState5 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(\"\"),\n      _useState6 = _slicedToArray(_useState5, 2),\n      errorMsg = _useState6[0],\n      setErrorMsg = _useState6[1];\n\n  var usernameInputChange = function usernameInputChange(e) {\n    setUsername(e.target.value);\n  };\n\n  var pwdInputChange = function pwdInputChange(e) {\n    setPwd(e.target.value);\n  };\n\n  var validateInput = function validateInput() {\n    return username !== \"\" && pwd !== \"\" ? true : false;\n  };\n\n  var onLoginClick = function onLoginClick() {\n    if (validateInput()) {\n      axios__WEBPACK_IMPORTED_MODULE_0___default()({\n        method: \"post\",\n        url: \"/api/login\",\n        headers: {\n          \"Content-Type\": \"application/json\"\n        },\n        data: {\n          username: username,\n          password: pwd\n        }\n      }).then(function (response) {\n        setErrorMsg(\"\");\n        saveUserData(response.data.token, response.data.user.username, response.data.user.pk);\n        return response.data.token;\n      })[\"catch\"](function (error) {\n        if (error.response) {\n          if (error.response.data[\"non_field_errors\"][0] === \"Unable to log in with provided credentials.\") setErrorMsg(\"아이디 혹은 비밀번호가 바르지 않습니다.\");else if (error.response.data[\"non_field_errors\"][0] === \"E-mail is not verified.\") setErrorMsg(\"이메일 인증 후 로그인 해주세요.\");else {\n            setErrorMsg(\"로그인에 실패했습니다. 다시 시도해주세요\");\n          }\n        }\n      });\n    } else {\n      setErrorMsg(\"아이디와 비밀번호를 입력해주세요\");\n    }\n  };\n\n  var handleKeyPress = function handleKeyPress(e) {\n    if (e.key === \"Enter\") {\n      onLoginClick();\n    }\n  };\n\n  return currUser ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(react_router__WEBPACK_IMPORTED_MODULE_6__.Redirect, {\n    to: \"/\"\n  }) : /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_styles_globalStyles__WEBPACK_IMPORTED_MODULE_3__.Background, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_Login_elements__WEBPACK_IMPORTED_MODULE_4__.LoginSection, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_Login_elements__WEBPACK_IMPORTED_MODULE_4__.LoginForm, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_Components_csrftoken__WEBPACK_IMPORTED_MODULE_2__[\"default\"], null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_Login_elements__WEBPACK_IMPORTED_MODULE_4__.LoginLegend, null, \"\\uB85C\\uADF8\\uC778\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_Login_elements__WEBPACK_IMPORTED_MODULE_4__.LoginFieldset, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_Login_elements__WEBPACK_IMPORTED_MODULE_4__.InputContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_Login_elements__WEBPACK_IMPORTED_MODULE_4__.LoginInput, {\n    type: \"text\",\n    name: \"username\",\n    placeholder: \"\\uC544\\uC774\\uB514\",\n    onChange: function onChange(e) {\n      usernameInputChange(e);\n    },\n    onKeyPress: handleKeyPress\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_Login_elements__WEBPACK_IMPORTED_MODULE_4__.InputContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_Login_elements__WEBPACK_IMPORTED_MODULE_4__.LoginInput, {\n    type: \"password\",\n    name: \"password\",\n    placeholder: \"\\uBE44\\uBC00\\uBC88\\uD638\",\n    onChange: function onChange(e) {\n      pwdInputChange(e);\n    },\n    onKeyPress: handleKeyPress\n  }))), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_Login_elements__WEBPACK_IMPORTED_MODULE_4__.LoginBtnContainer, null, errorMsg !== \"\" ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_Login_elements__WEBPACK_IMPORTED_MODULE_4__.ErrorMessage, null, errorMsg) : \"\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_Login_elements__WEBPACK_IMPORTED_MODULE_4__.LoginBtn, {\n    onClick: function onClick() {\n      return onLoginClick();\n    }\n  }, \"\\uB85C\\uADF8\\uC778\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_Login_elements__WEBPACK_IMPORTED_MODULE_4__.ToRegisterLink, {\n    to: \"/register\"\n  }, \"\\uD68C\\uC6D0\\uAC00\\uC785\\uD558\\uAE30\")))));\n};\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (Login);\n\n//# sourceURL=webpack://frontend/./src/Containers/Login/Login.js?");
+
+/***/ }),
+
+/***/ "./src/Containers/MusicCardGrid/MusicCardGrid.elements.js":
+/*!****************************************************************!*\
+  !*** ./src/Containers/MusicCardGrid/MusicCardGrid.elements.js ***!
+  \****************************************************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"MusicListContainer\": () => (/* binding */ MusicListContainer),\n/* harmony export */   \"MusicCardWrapper\": () => (/* binding */ MusicCardWrapper)\n/* harmony export */ });\n/* harmony import */ var styled_components__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! styled-components */ \"./node_modules/styled-components/dist/styled-components.browser.esm.js\");\n/* harmony import */ var _styles_Colors__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../styles/Colors */ \"./src/styles/Colors.js\");\n/* harmony import */ var _styles_globalStyles__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../styles/globalStyles */ \"./src/styles/globalStyles.js\");\nvar _templateObject, _templateObject2;\n\nfunction _taggedTemplateLiteral(strings, raw) { if (!raw) { raw = strings.slice(0); } return Object.freeze(Object.defineProperties(strings, { raw: { value: Object.freeze(raw) } })); }\n\n\n\n\nvar MusicListContainer = styled_components__WEBPACK_IMPORTED_MODULE_2__[\"default\"].div(_templateObject || (_templateObject = _taggedTemplateLiteral([\"\\n  width: 100%;\\n  display: grid;\\n  grid-template-columns: repeat(3, 1fr);\\n  justify-items: center;\\n  gap: 5%;\\n  margin: 5% auto;\\n\"])));\nvar MusicCardWrapper = styled_components__WEBPACK_IMPORTED_MODULE_2__[\"default\"].div(_templateObject2 || (_templateObject2 = _taggedTemplateLiteral([\"\\n  cursor: pointer;\\n  width: 18rem;\\n  height: 7.5rem;\\n  display: flex;\\n  align-items: center;\\n  border: 1px solid \", \";\\n  border-radius: 5px;\\n  box-shadow: \", \";\\n  \", \"\\n\"])), _styles_Colors__WEBPACK_IMPORTED_MODULE_0__.cardStroke, _styles_globalStyles__WEBPACK_IMPORTED_MODULE_1__.CardShadow, function (p) {\n  return p.open === true ? \"border: 1px solid \".concat(_styles_Colors__WEBPACK_IMPORTED_MODULE_0__.iiPurple, \";\") : \"border: 1px solid \".concat(_styles_Colors__WEBPACK_IMPORTED_MODULE_0__.cardStroke, \";\");\n});\n\n//# sourceURL=webpack://frontend/./src/Containers/MusicCardGrid/MusicCardGrid.elements.js?");
+
+/***/ }),
+
+/***/ "./src/Containers/MusicCardGrid/MusicCardGrid.js":
+/*!*******************************************************!*\
+  !*** ./src/Containers/MusicCardGrid/MusicCardGrid.js ***!
+  \*******************************************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var _MusicCardGrid_elements__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./MusicCardGrid.elements */ \"./src/Containers/MusicCardGrid/MusicCardGrid.elements.js\");\n/* harmony import */ var _Components__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../Components */ \"./src/Components/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! axios */ \"./node_modules/axios/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_3___default = /*#__PURE__*/__webpack_require__.n(axios__WEBPACK_IMPORTED_MODULE_3__);\nfunction _slicedToArray(arr, i) { return _arrayWithHoles(arr) || _iterableToArrayLimit(arr, i) || _unsupportedIterableToArray(arr, i) || _nonIterableRest(); }\n\nfunction _nonIterableRest() { throw new TypeError(\"Invalid attempt to destructure non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); }\n\nfunction _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === \"string\") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === \"Object\" && o.constructor) n = o.constructor.name; if (n === \"Map\" || n === \"Set\") return Array.from(o); if (n === \"Arguments\" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }\n\nfunction _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }\n\nfunction _iterableToArrayLimit(arr, i) { var _i = arr == null ? null : typeof Symbol !== \"undefined\" && arr[Symbol.iterator] || arr[\"@@iterator\"]; if (_i == null) return; var _arr = []; var _n = true; var _d = false; var _s, _e; try { for (_i = _i.call(arr); !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i[\"return\"] != null) _i[\"return\"](); } finally { if (_d) throw _e; } } return _arr; }\n\nfunction _arrayWithHoles(arr) { if (Array.isArray(arr)) return arr; }\n\n\n\n\n\n\nvar MusicCardGrid = function MusicCardGrid(_ref) {\n  var currUser = _ref.currUser,\n      token = _ref.token,\n      userId = _ref.userId,\n      list = _ref.list;\n\n  var _useState = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(null),\n      _useState2 = _slicedToArray(_useState, 2),\n      content = _useState2[0],\n      setContent = _useState2[1];\n\n  var _useState3 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(0),\n      _useState4 = _slicedToArray(_useState3, 2),\n      openCard = _useState4[0],\n      setOpenCard = _useState4[1];\n\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    setContent();\n\n    if (list && list.length) {\n      setContent(list);\n    }\n  }, [list]);\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    if (!list) {\n      axios__WEBPACK_IMPORTED_MODULE_3___default()({\n        method: \"get\",\n        url: \"/api/recentmusic/6\",\n        headers: {\n          \"Content-Type\": \"application/json\"\n        }\n      }).then(function (response) {\n        if (response.status > 400) {\n          setPlaceholder(\"Something went wrong!\");\n        }\n\n        return response.data;\n      }).then(function (data) {\n        setContent(data[\"music list\"]);\n      });\n    }\n  }, []);\n\n  var handleClick = function handleClick(clickedCard) {\n    setOpenCard(clickedCard);\n  };\n\n  var MusicCardList = content ? content.map(function (c, i) {\n    return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_MusicCardGrid_elements__WEBPACK_IMPORTED_MODULE_1__.MusicCardWrapper, {\n      key: i,\n      onclick: function onclick() {\n        return handleClick(i);\n      },\n      open: openCard === i\n    }, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_2__.MusicCard, {\n      title: c[0],\n      artist: c[1],\n      cover: c[2],\n      source: c[3],\n      link: c[4],\n      cardIndex: i,\n      handleClick: handleClick,\n      open: openCard === i\n    }));\n  }) : [];\n  return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(react__WEBPACK_IMPORTED_MODULE_0__.Fragment, null, content ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(react__WEBPACK_IMPORTED_MODULE_0__.Fragment, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_MusicCardGrid_elements__WEBPACK_IMPORTED_MODULE_1__.MusicListContainer, null, MusicCardList), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_2__.MusicCardOpened, {\n    currUser: currUser,\n    token: token,\n    userId: userId,\n    cardIndex: openCard,\n    title: content[openCard] ? content[openCard][0] : \"\",\n    artist: content[openCard] ? content[openCard][1] : \"\",\n    cover: content[openCard] ? content[openCard][2] : \"\",\n    source: content[openCard] ? content[openCard][3] : \"\",\n    link: content[openCard] ? content[openCard][4] : \"\"\n  })) : \"\");\n};\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (MusicCardGrid);\n\n//# sourceURL=webpack://frontend/./src/Containers/MusicCardGrid/MusicCardGrid.js?");
 
 /***/ }),
 
@@ -932,6 +910,28 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 
 /***/ }),
 
+/***/ "./src/Containers/QuestionCardGrid/QuestionCardGrid.elements.js":
+/*!**********************************************************************!*\
+  !*** ./src/Containers/QuestionCardGrid/QuestionCardGrid.elements.js ***!
+  \**********************************************************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"QuestionGridContainer\": () => (/* binding */ QuestionGridContainer),\n/* harmony export */   \"QuestionCardWrapper\": () => (/* binding */ QuestionCardWrapper)\n/* harmony export */ });\n/* harmony import */ var styled_components__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! styled-components */ \"./node_modules/styled-components/dist/styled-components.browser.esm.js\");\n/* harmony import */ var _styles_Colors__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../styles/Colors */ \"./src/styles/Colors.js\");\n/* harmony import */ var _styles_globalStyles__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../styles/globalStyles */ \"./src/styles/globalStyles.js\");\nvar _templateObject, _templateObject2;\n\nfunction _taggedTemplateLiteral(strings, raw) { if (!raw) { raw = strings.slice(0); } return Object.freeze(Object.defineProperties(strings, { raw: { value: Object.freeze(raw) } })); }\n\n\n\n\nvar QuestionGridContainer = styled_components__WEBPACK_IMPORTED_MODULE_2__[\"default\"].div(_templateObject || (_templateObject = _taggedTemplateLiteral([\"\\n  display: grid;\\n  grid-template-columns: repeat(3, 29vw);\\n  grid-gap: 1.5vw;\\n  margin: 3% auto;\\n  width: 100%;\\n\\n  @media screen and (min-width: 1920px) {\\n    grid-template-columns: repeat(3, 19vw);\\n  }\\n\"])));\nvar QuestionCardWrapper = styled_components__WEBPACK_IMPORTED_MODULE_2__[\"default\"].div(_templateObject2 || (_templateObject2 = _taggedTemplateLiteral([\"\\n  text-align: center;\\n\"])));\n\n//# sourceURL=webpack://frontend/./src/Containers/QuestionCardGrid/QuestionCardGrid.elements.js?");
+
+/***/ }),
+
+/***/ "./src/Containers/QuestionCardGrid/QuestionCardGrid.js":
+/*!*************************************************************!*\
+  !*** ./src/Containers/QuestionCardGrid/QuestionCardGrid.js ***!
+  \*************************************************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var _QuestionCardGrid_elements__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./QuestionCardGrid.elements */ \"./src/Containers/QuestionCardGrid/QuestionCardGrid.elements.js\");\n/* harmony import */ var _Components__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../Components */ \"./src/Components/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! axios */ \"./node_modules/axios/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_3___default = /*#__PURE__*/__webpack_require__.n(axios__WEBPACK_IMPORTED_MODULE_3__);\nfunction _slicedToArray(arr, i) { return _arrayWithHoles(arr) || _iterableToArrayLimit(arr, i) || _unsupportedIterableToArray(arr, i) || _nonIterableRest(); }\n\nfunction _nonIterableRest() { throw new TypeError(\"Invalid attempt to destructure non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); }\n\nfunction _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === \"string\") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === \"Object\" && o.constructor) n = o.constructor.name; if (n === \"Map\" || n === \"Set\") return Array.from(o); if (n === \"Arguments\" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }\n\nfunction _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }\n\nfunction _iterableToArrayLimit(arr, i) { var _i = arr == null ? null : typeof Symbol !== \"undefined\" && arr[Symbol.iterator] || arr[\"@@iterator\"]; if (_i == null) return; var _arr = []; var _n = true; var _d = false; var _s, _e; try { for (_i = _i.call(arr); !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i[\"return\"] != null) _i[\"return\"](); } finally { if (_d) throw _e; } } return _arr; }\n\nfunction _arrayWithHoles(arr) { if (Array.isArray(arr)) return arr; }\n\n\n\n\n\n\nvar QuestionCardGrid = function QuestionCardGrid(_ref) {\n  var currUser = _ref.currUser,\n      token = _ref.token,\n      userId = _ref.userId,\n      list = _ref.list;\n\n  var _useState = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(),\n      _useState2 = _slicedToArray(_useState, 2),\n      content = _useState2[0],\n      setContent = _useState2[1];\n\n  var _useState3 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(-1),\n      _useState4 = _slicedToArray(_useState3, 2),\n      openCard = _useState4[0],\n      setOpenCard = _useState4[1];\n\n  var _useState5 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(false),\n      _useState6 = _slicedToArray(_useState5, 2),\n      isOpened = _useState6[0],\n      setIsOpened = _useState6[1];\n\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    setContent();\n\n    if (list && list.length) {\n      setContent(list);\n    }\n  }, [list]);\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    if (!list) {\n      axios__WEBPACK_IMPORTED_MODULE_3___default()({\n        method: \"get\",\n        url: \"/api/question\",\n        headers: {\n          \"Content-Type\": \"application/json\"\n        }\n      }).then(function (response) {\n        if (response.status > 400) {\n          setPlaceholder(\"Something went wrong!\");\n        }\n\n        return response.data;\n      }).then(function (data) {\n        setContent(data);\n      });\n    }\n  }, []);\n\n  var handleClick = function handleClick(clickedCard) {\n    if (clickedCard === openCard) {\n      setOpenCard(-1);\n      setIsOpened(false);\n    } else {\n      //다른 카드 클릭\n      setOpenCard(clickedCard);\n      setIsOpened(true);\n    }\n  };\n\n  var QuestionCardList = content ? content.map(function (c, i) {\n    return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_QuestionCardGrid_elements__WEBPACK_IMPORTED_MODULE_1__.QuestionCardWrapper, {\n      key: i\n    }, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_2__.QuestionCard, {\n      question: c.question_content,\n      date: c.released_date,\n      cardIndex: i,\n      handleClick: handleClick,\n      open: isOpened && openCard === i\n    }), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_2__.QuestionOpened, {\n      currUser: currUser,\n      token: token,\n      userId: userId,\n      questionId: c.id,\n      question: c.question_content,\n      detail: c.explain,\n      cardIndex: i,\n      open: isOpened && openCard === i\n    }));\n  }) : [];\n  return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_QuestionCardGrid_elements__WEBPACK_IMPORTED_MODULE_1__.QuestionGridContainer, null, QuestionCardList);\n};\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (QuestionCardGrid);\n\n//# sourceURL=webpack://frontend/./src/Containers/QuestionCardGrid/QuestionCardGrid.js?");
+
+/***/ }),
+
 /***/ "./src/Containers/Search/Search.elements.js":
 /*!**************************************************!*\
   !*** ./src/Containers/Search/Search.elements.js ***!
@@ -950,7 +950,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! axios */ \"./node_modules/axios/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(axios__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var ___WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! .. */ \"./src/Containers/index.js\");\n/* harmony import */ var _Search_elements__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./Search.elements */ \"./src/Containers/Search/Search.elements.js\");\n/* harmony import */ var _styles_globalStyles__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ../../styles/globalStyles */ \"./src/styles/globalStyles.js\");\n/* harmony import */ var _Components__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ../../Components */ \"./src/Components/index.js\");\nfunction _slicedToArray(arr, i) { return _arrayWithHoles(arr) || _iterableToArrayLimit(arr, i) || _unsupportedIterableToArray(arr, i) || _nonIterableRest(); }\n\nfunction _nonIterableRest() { throw new TypeError(\"Invalid attempt to destructure non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); }\n\nfunction _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === \"string\") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === \"Object\" && o.constructor) n = o.constructor.name; if (n === \"Map\" || n === \"Set\") return Array.from(o); if (n === \"Arguments\" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }\n\nfunction _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }\n\nfunction _iterableToArrayLimit(arr, i) { var _i = arr == null ? null : typeof Symbol !== \"undefined\" && arr[Symbol.iterator] || arr[\"@@iterator\"]; if (_i == null) return; var _arr = []; var _n = true; var _d = false; var _s, _e; try { for (_i = _i.call(arr); !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i[\"return\"] != null) _i[\"return\"](); } finally { if (_d) throw _e; } } return _arr; }\n\nfunction _arrayWithHoles(arr) { if (Array.isArray(arr)) return arr; }\n\n\n\n\n\n\n\n\nvar Search = function Search(_ref) {\n  var currUser = _ref.currUser,\n      token = _ref.token,\n      userId = _ref.userId,\n      word = _ref.word;\n\n  var _useState = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)([]),\n      _useState2 = _slicedToArray(_useState, 2),\n      postResult = _useState2[0],\n      setPostResult = _useState2[1];\n\n  var _useState3 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)([]),\n      _useState4 = _slicedToArray(_useState3, 2),\n      questionResult = _useState4[0],\n      setQuestionResult = _useState4[1];\n\n  var _useState5 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)([]),\n      _useState6 = _slicedToArray(_useState5, 2),\n      musicResult = _useState6[0],\n      setMusicResult = _useState6[1];\n\n  var _useState7 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(word),\n      _useState8 = _slicedToArray(_useState7, 2),\n      keyword = _useState8[0],\n      setKeyword = _useState8[1];\n\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    setPostResult([]);\n    setQuestionResult([]);\n    setMusicResult([]);\n    axios__WEBPACK_IMPORTED_MODULE_1___default()({\n      method: \"get\",\n      url: \"/api/search/\".concat(keyword),\n      headers: {\n        \"Content-Type\": \"application/json\"\n      }\n    }).then(function (response) {\n      if (response.status > 400) {\n        setPlaceholder(\"Something went wrong!\");\n      }\n\n      return response.data;\n    }).then(function (data) {\n      setPostResult(data.post);\n      setQuestionResult(data.question);\n      setMusicResult(data.music);\n      return data;\n    });\n  }, [keyword]);\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    setPostResult([]);\n    setQuestionResult([]);\n    setMusicResult([]);\n    setKeyword(word);\n    axios__WEBPACK_IMPORTED_MODULE_1___default()({\n      method: \"get\",\n      url: \"/api/search/\".concat(word),\n      headers: {\n        \"Content-Type\": \"application/json\"\n      }\n    }).then(function (response) {\n      if (response.status > 400) {\n        setPlaceholder(\"Something went wrong!\");\n      }\n\n      return response.data;\n    }).then(function (data) {\n      setPostResult(data.post);\n      setQuestionResult(data.question);\n      setMusicResult(data.music);\n      return data;\n    });\n  }, [word]);\n  return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(react__WEBPACK_IMPORTED_MODULE_0__.Fragment, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_styles_globalStyles__WEBPACK_IMPORTED_MODULE_4__.Background, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_styles_globalStyles__WEBPACK_IMPORTED_MODULE_4__.PageSection, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.SearchBarContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.SearchBar, {\n    value: keyword,\n    placeholder: keyword,\n    onChange: function onChange(e) {\n      setKeyword(e.target.value);\n    }\n  }), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.SearchButton, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"i\", {\n    className: \"fas fa-search\"\n  }))), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.PostResultContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.ResultTitle, null, \"\\uAE00 \\uAC80\\uC0C9 \\uACB0\\uACFC\"), postResult.length === 0 ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.NoResultWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.NoResult, null, \"\\uAC80\\uC0C9\\uACB0\\uACFC\\uAC00 \\uC5C6\\uC2B5\\uB2C8\\uB2E4\")) : /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(___WEBPACK_IMPORTED_MODULE_2__.PostCardList, {\n    currUser: currUser,\n    token: token,\n    userId: userId,\n    list: postResult\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.QuestionResultContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.ResultTitle, null, \"\\uC9C8\\uBB38 \\uAC80\\uC0C9 \\uACB0\\uACFC\"), questionResult.length === 0 ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.NoResultWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.NoResult, null, \"\\uAC80\\uC0C9\\uACB0\\uACFC\\uAC00 \\uC5C6\\uC2B5\\uB2C8\\uB2E4\")) : /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_5__.QuestionCardGrid, {\n    currUser: currUser,\n    token: token,\n    userId: userId,\n    list: questionResult\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.MusicResultContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.ResultTitle, null, \"\\uC74C\\uC545 \\uAC80\\uC0C9 \\uACB0\\uACFC\"), musicResult.length === 0 ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.NoResultWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.NoResult, null, \"\\uAC80\\uC0C9\\uACB0\\uACFC\\uAC00 \\uC5C6\\uC2B5\\uB2C8\\uB2E4\")) : /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_5__.MusicCardGrid, {\n    currUser: currUser,\n    token: token,\n    userId: userId,\n    list: musicResult\n  })))));\n};\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (Search);\n\n//# sourceURL=webpack://frontend/./src/Containers/Search/Search.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! axios */ \"./node_modules/axios/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(axios__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var ___WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../ */ \"./src/Containers/index.js\");\n/* harmony import */ var _Search_elements__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./Search.elements */ \"./src/Containers/Search/Search.elements.js\");\n/* harmony import */ var _styles_globalStyles__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ../../styles/globalStyles */ \"./src/styles/globalStyles.js\");\nfunction _slicedToArray(arr, i) { return _arrayWithHoles(arr) || _iterableToArrayLimit(arr, i) || _unsupportedIterableToArray(arr, i) || _nonIterableRest(); }\n\nfunction _nonIterableRest() { throw new TypeError(\"Invalid attempt to destructure non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); }\n\nfunction _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === \"string\") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === \"Object\" && o.constructor) n = o.constructor.name; if (n === \"Map\" || n === \"Set\") return Array.from(o); if (n === \"Arguments\" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }\n\nfunction _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }\n\nfunction _iterableToArrayLimit(arr, i) { var _i = arr == null ? null : typeof Symbol !== \"undefined\" && arr[Symbol.iterator] || arr[\"@@iterator\"]; if (_i == null) return; var _arr = []; var _n = true; var _d = false; var _s, _e; try { for (_i = _i.call(arr); !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i[\"return\"] != null) _i[\"return\"](); } finally { if (_d) throw _e; } } return _arr; }\n\nfunction _arrayWithHoles(arr) { if (Array.isArray(arr)) return arr; }\n\n\n\n\n\n\n\n\nvar Search = function Search(_ref) {\n  var currUser = _ref.currUser,\n      token = _ref.token,\n      userId = _ref.userId,\n      word = _ref.word;\n\n  var _useState = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)([]),\n      _useState2 = _slicedToArray(_useState, 2),\n      postResult = _useState2[0],\n      setPostResult = _useState2[1];\n\n  var _useState3 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)([]),\n      _useState4 = _slicedToArray(_useState3, 2),\n      questionResult = _useState4[0],\n      setQuestionResult = _useState4[1];\n\n  var _useState5 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)([]),\n      _useState6 = _slicedToArray(_useState5, 2),\n      musicResult = _useState6[0],\n      setMusicResult = _useState6[1];\n\n  var _useState7 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(word),\n      _useState8 = _slicedToArray(_useState7, 2),\n      keyword = _useState8[0],\n      setKeyword = _useState8[1];\n\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    setPostResult([]);\n    setQuestionResult([]);\n    setMusicResult([]);\n    axios__WEBPACK_IMPORTED_MODULE_1___default()({\n      method: \"get\",\n      url: \"/api/search/\".concat(keyword),\n      headers: {\n        \"Content-Type\": \"application/json\"\n      }\n    }).then(function (response) {\n      if (response.status > 400) {\n        setPlaceholder(\"Something went wrong!\");\n      }\n\n      return response.data;\n    }).then(function (data) {\n      setPostResult(data.post);\n      setQuestionResult(data.question);\n      setMusicResult(data.music);\n      return data;\n    });\n  }, [keyword]);\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    setPostResult([]);\n    setQuestionResult([]);\n    setMusicResult([]);\n    setKeyword(word);\n    axios__WEBPACK_IMPORTED_MODULE_1___default()({\n      method: \"get\",\n      url: \"/api/search/\".concat(word),\n      headers: {\n        \"Content-Type\": \"application/json\"\n      }\n    }).then(function (response) {\n      if (response.status > 400) {\n        setPlaceholder(\"Something went wrong!\");\n      }\n\n      return response.data;\n    }).then(function (data) {\n      setPostResult(data.post);\n      setQuestionResult(data.question);\n      setMusicResult(data.music);\n      return data;\n    });\n  }, [word]);\n  return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(react__WEBPACK_IMPORTED_MODULE_0__.Fragment, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_styles_globalStyles__WEBPACK_IMPORTED_MODULE_4__.Background, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_styles_globalStyles__WEBPACK_IMPORTED_MODULE_4__.PageSection, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.SearchBarContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.SearchBar, {\n    value: keyword,\n    placeholder: keyword,\n    onChange: function onChange(e) {\n      setKeyword(e.target.value);\n    }\n  }), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.SearchButton, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"i\", {\n    className: \"fas fa-search\"\n  }))), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.PostResultContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.ResultTitle, null, \"\\uAE00 \\uAC80\\uC0C9 \\uACB0\\uACFC\"), postResult.length === 0 ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.NoResultWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.NoResult, null, \"\\uAC80\\uC0C9\\uACB0\\uACFC\\uAC00 \\uC5C6\\uC2B5\\uB2C8\\uB2E4\")) : /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(___WEBPACK_IMPORTED_MODULE_2__.PostCardList, {\n    currUser: currUser,\n    token: token,\n    userId: userId,\n    list: postResult\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.QuestionResultContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.ResultTitle, null, \"\\uC9C8\\uBB38 \\uAC80\\uC0C9 \\uACB0\\uACFC\"), questionResult.length === 0 ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.NoResultWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.NoResult, null, \"\\uAC80\\uC0C9\\uACB0\\uACFC\\uAC00 \\uC5C6\\uC2B5\\uB2C8\\uB2E4\")) : /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(___WEBPACK_IMPORTED_MODULE_2__.QuestionCardGrid, {\n    currUser: currUser,\n    token: token,\n    userId: userId,\n    list: questionResult\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.MusicResultContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.ResultTitle, null, \"\\uC74C\\uC545 \\uAC80\\uC0C9 \\uACB0\\uACFC\"), musicResult.length === 0 ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.NoResultWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Search_elements__WEBPACK_IMPORTED_MODULE_3__.NoResult, null, \"\\uAC80\\uC0C9\\uACB0\\uACFC\\uAC00 \\uC5C6\\uC2B5\\uB2C8\\uB2E4\")) : /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(___WEBPACK_IMPORTED_MODULE_2__.MusicCardGrid, {\n    currUser: currUser,\n    token: token,\n    userId: userId,\n    list: musicResult\n  })))));\n};\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (Search);\n\n//# sourceURL=webpack://frontend/./src/Containers/Search/Search.js?");
 
 /***/ }),
 
@@ -972,7 +972,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! axios */ \"./node_modules/axios/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(axios__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var _Components_csrftoken__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../Components/csrftoken */ \"./src/Components/csrftoken.js\");\n/* harmony import */ var _styles_globalStyles__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ../../styles/globalStyles */ \"./src/styles/globalStyles.js\");\n/* harmony import */ var _SignUp_elements__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./SignUp.elements */ \"./src/Containers/SignUp/SignUp.elements.js\");\nfunction _slicedToArray(arr, i) { return _arrayWithHoles(arr) || _iterableToArrayLimit(arr, i) || _unsupportedIterableToArray(arr, i) || _nonIterableRest(); }\n\nfunction _nonIterableRest() { throw new TypeError(\"Invalid attempt to destructure non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); }\n\nfunction _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === \"string\") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === \"Object\" && o.constructor) n = o.constructor.name; if (n === \"Map\" || n === \"Set\") return Array.from(o); if (n === \"Arguments\" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }\n\nfunction _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }\n\nfunction _iterableToArrayLimit(arr, i) { var _i = arr == null ? null : typeof Symbol !== \"undefined\" && arr[Symbol.iterator] || arr[\"@@iterator\"]; if (_i == null) return; var _arr = []; var _n = true; var _d = false; var _s, _e; try { for (_i = _i.call(arr); !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i[\"return\"] != null) _i[\"return\"](); } finally { if (_d) throw _e; } } return _arr; }\n\nfunction _arrayWithHoles(arr) { if (Array.isArray(arr)) return arr; }\n\n\n\n\n\n\n\n\nvar SignUp = function SignUp() {\n  var _useState = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(\"\"),\n      _useState2 = _slicedToArray(_useState, 2),\n      username = _useState2[0],\n      setUsername = _useState2[1];\n\n  var _useState3 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(\"\"),\n      _useState4 = _slicedToArray(_useState3, 2),\n      nickname = _useState4[0],\n      setNickname = _useState4[1];\n\n  var _useState5 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(\"\"),\n      _useState6 = _slicedToArray(_useState5, 2),\n      email = _useState6[0],\n      setEmail = _useState6[1];\n\n  var _useState7 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(\"\"),\n      _useState8 = _slicedToArray(_useState7, 2),\n      pwd = _useState8[0],\n      setPwd = _useState8[1];\n\n  var _useState9 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(\"\"),\n      _useState10 = _slicedToArray(_useState9, 2),\n      pwdConfirm = _useState10[0],\n      setConfirm = _useState10[1];\n\n  var _useState11 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(false),\n      _useState12 = _slicedToArray(_useState11, 2),\n      submit = _useState12[0],\n      setSubmit = _useState12[1];\n\n  var _useState13 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(false),\n      _useState14 = _slicedToArray(_useState13, 2),\n      sent = _useState14[0],\n      setSent = _useState14[1];\n\n  var _useState15 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(),\n      _useState16 = _slicedToArray(_useState15, 2),\n      pwdMatch = _useState16[0],\n      setPwdMatch = _useState16[1];\n\n  var _useState17 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(),\n      _useState18 = _slicedToArray(_useState17, 2),\n      usernameChecked = _useState18[0],\n      setUsernameCheck = _useState18[1];\n\n  var _useState19 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(),\n      _useState20 = _slicedToArray(_useState19, 2),\n      nicknameChecked = _useState20[0],\n      setNicknameCheck = _useState20[1];\n\n  var _useState21 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(false),\n      _useState22 = _slicedToArray(_useState21, 2),\n      requestReview = _useState22[0],\n      setRequestReview = _useState22[1];\n\n  var onRegisterClick = function onRegisterClick() {\n    setSubmit(true);\n\n    if (validateInput()) {\n      axios__WEBPACK_IMPORTED_MODULE_0___default()({\n        method: \"post\",\n        url: \"/api/signup\",\n        headers: {\n          \"Content-Type\": \"application/json\"\n        },\n        data: {\n          username: username,\n          nickname: nickname,\n          email: email,\n          password1: pwd,\n          password2: pwdConfirm\n        }\n      })[\"catch\"](function (err) {\n        return err.response.status === 400 ? setRequestReview(\"비밀번호가 보안에 취약합니다. 다른 비밀번호를 입력해주세요.\") : err.response;\n      }).then(function (response) {\n        if (response.status < 400) {\n          setSent(true);\n        }\n      });\n    }\n  };\n\n  var validateInput = function validateInput() {\n    if (username === \"\" || email === \"\" || pwd === \"\" || pwdConfirm === \"\") {\n      setRequestReview(\"모든 영역은 필수 입력입니다.\");\n      return false;\n    } else if (!usernameChecked || !nicknameChecked) {\n      setRequestReview(\"아이디와 필명의 중복 확인이 필요합니다.\");\n      return false;\n    } else if (!pwdMatch) {\n      setRequestReview(\"비밀번호가 일치하지 않습니다.\");\n      return false;\n    } else if (pwd.length < 8) {\n      setRequestReview(\"비밀번호가 너무 짧습니다.\");\n      return false;\n    } else if (pwd.match(/^\\d+$/)) {\n      setRequestReview(\"비밀번호가 숫자로만 이루어져 있습니다.\");\n      return false;\n    }\n\n    return true;\n  };\n\n  var checkDuplicate = function checkDuplicate(type, target, setFunc) {\n    if (target === \"\") setFunc(false);else {\n      axios__WEBPACK_IMPORTED_MODULE_0___default()({\n        method: \"post\",\n        url: \"/api/\".concat(type, \"-check\"),\n        headers: {\n          \"Content-Type\": \"application/json\"\n        },\n        data: {\n          username: \"\".concat(target),\n          nickname: \"\".concat(target)\n        }\n      })[\"catch\"](function (error) {\n        if (error.response.status === 409) {\n          setFunc(false);\n          return error.response;\n        }\n      }).then(function (response) {\n        if (response.status < 400) {\n          setFunc(true);\n        }\n      });\n    }\n  };\n\n  (0,react__WEBPACK_IMPORTED_MODULE_1__.useEffect)(function () {\n    if (pwdConfirm === \"\") setPwdMatch();else if (pwd === pwdConfirm) setPwdMatch(true);else setPwdMatch(false);\n  }, [pwdConfirm]);\n  return sent ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_styles_globalStyles__WEBPACK_IMPORTED_MODULE_3__.Background, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.AfterSent, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.SentMessage, null, \"\\uC785\\uB825\\uD574\\uC8FC\\uC2E0 \\uC774\\uBA54\\uC77C\\uB85C \\uC778\\uC99D \\uB9C1\\uD06C\\uB97C \\uBCF4\\uB0B4\\uB4DC\\uB838\\uC2B5\\uB2C8\\uB2E4. \\uC778\\uC99D \\uD6C4 \\uB85C\\uADF8\\uC778 \\uAC00\\uB2A5\\uD569\\uB2C8\\uB2E4.\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.ToLoginLink, {\n    to: \"/login\"\n  }, \"\\uB85C\\uADF8\\uC778 \\uD558\\uAE30\"))) : /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_styles_globalStyles__WEBPACK_IMPORTED_MODULE_3__.Background, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterSection, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterForm, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_Components_csrftoken__WEBPACK_IMPORTED_MODULE_2__[\"default\"], null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterLegend, null, \"\\uD68C\\uC6D0\\uAC00\\uC785\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterFieldset, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputLabel, {\n    htmlFor: \"username\"\n  }, \"\\uC544\\uC774\\uB514\", \" \", usernameChecked === false ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.DuplicateMessage, null, \"\\uD574\\uB2F9 \\uC544\\uC774\\uB514\\uB294 \\uC0AC\\uC6A9\\uD560 \\uC218 \\uC5C6\\uC2B5\\uB2C8\\uB2E4.\") : \"\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterInput, {\n    type: \"text\",\n    name: \"username\",\n    placeholder: \"\\uACF5\\uBC31\\uC5C6\\uC774 \\uC601\\uBB38, \\uC22B\\uC790 \\uD3EC\\uD568 6-12\\uC790\",\n    duplicateChecked: usernameChecked,\n    onChange: function onChange(e) {\n      setUsername(e.target.value);\n      setUsernameCheck();\n      setRequestReview(\"\");\n    }\n  }), !usernameChecked ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.DuplicateCheckButton, {\n    onClick: function onClick() {\n      checkDuplicate(\"username\", username, setUsernameCheck);\n    }\n  }, \"\\uC911\\uBCF5 \\uD655\\uC778\") : /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.DuplicateConfirm, null, \"\\uC0AC\\uC6A9 \\uAC00\\uB2A5\\uD569\\uB2C8\\uB2E4\"))), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputLabel, {\n    htmlFor: \"nickname\"\n  }, \"\\uD544\\uBA85\", \" \", nicknameChecked === false ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.DuplicateMessage, null, \"\\uD574\\uB2F9 \\uD544\\uBA85\\uC740 \\uC0AC\\uC6A9\\uD560 \\uC218 \\uC5C6\\uC2B5\\uB2C8\\uB2E4.\") : \"\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterInput, {\n    type: \"text\",\n    name: \"nickname\",\n    placeholder: \"\\uAE00\\uC744 \\uC791\\uC131 \\uC2DC \\uD45C\\uC2DC\\uB418\\uB294 \\uC774\\uB984\",\n    duplicateChecked: nicknameChecked,\n    onChange: function onChange(e) {\n      setNickname(e.target.value);\n      setNicknameCheck();\n      setRequestReview(\"\");\n    }\n  }), !nicknameChecked ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.DuplicateCheckButton, {\n    onClick: function onClick() {\n      checkDuplicate(\"nickname\", nickname, setNicknameCheck);\n    }\n  }, \"\\uC911\\uBCF5 \\uD655\\uC778\") : /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.DuplicateConfirm, null, \"\\uC0AC\\uC6A9 \\uAC00\\uB2A5\\uD569\\uB2C8\\uB2E4\"))), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputLabel, {\n    htmlFor: \"username\"\n  }, \"\\uC774\\uBA54\\uC77C\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterInput, {\n    type: \"email\",\n    name: \"email\",\n    placeholder: \"ex. example@iieum.com\",\n    onChange: function onChange(e) {\n      setEmail(e.target.value);\n      setRequestReview(\"\");\n    }\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputLabel, {\n    htmlFor: \"password\"\n  }, \"\\uBE44\\uBC00\\uBC88\\uD638\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterInput, {\n    type: \"password\",\n    name: \"password\",\n    placeholder: \"\\uACF5\\uBC31\\uC5C6\\uC774 \\uC601\\uBB38, \\uC22B\\uC790 \\uD3EC\\uD568 8-20\\uC790\",\n    duplicateChecked: pwdMatch,\n    onChange: function onChange(e) {\n      setPwd(e.target.value);\n      setRequestReview(\"\");\n    }\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputLabel, {\n    htmlFor: \"passwordConfirm\"\n  }, \"\\uBE44\\uBC00\\uBC88\\uD638 \\uD655\\uC778\", pwdMatch === false ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.DuplicateMessage, null, \"\\uBE44\\uBC00\\uBC88\\uD638\\uAC00 \\uC77C\\uCE58\\uD558\\uC9C0 \\uC54A\\uC2B5\\uB2C8\\uB2E4.\") : \"\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterInput, {\n    type: \"password\",\n    name: \"passwordConfirm\",\n    placeholder: \"\\uBE44\\uBC00\\uBC88\\uD638 \\uC7AC\\uD655\\uC778\",\n    duplicateChecked: pwdMatch,\n    onChange: function onChange(e) {\n      setConfirm(e.target.value);\n      setRequestReview(\"\");\n    }\n  }))), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterBtnContainer, null, requestReview ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(react__WEBPACK_IMPORTED_MODULE_1__.Fragment, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.DuplicateMessage, null, requestReview), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(\"br\", null)) : \"\", submit ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.SubmitMessage, null, \"\\uCC98\\uB9AC \\uC911\\uC785\\uB2C8\\uB2E4...\") : /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(react__WEBPACK_IMPORTED_MODULE_1__.Fragment, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterBtn, {\n    onClick: function onClick() {\n      return onRegisterClick();\n    }\n  }, \"\\uAC00\\uC785\\uD558\\uAE30\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.ToLoginLink, {\n    to: \"/login\"\n  }, \"\\uB85C\\uADF8\\uC778\\uD558\\uAE30\"))))));\n};\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (SignUp);\n\n//# sourceURL=webpack://frontend/./src/Containers/SignUp/SignUp.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! axios */ \"./node_modules/axios/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(axios__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var _Components_csrftoken__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../Components/csrftoken */ \"./src/Components/csrftoken.js\");\n/* harmony import */ var _styles_globalStyles__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ../../styles/globalStyles */ \"./src/styles/globalStyles.js\");\n/* harmony import */ var _SignUp_elements__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./SignUp.elements */ \"./src/Containers/SignUp/SignUp.elements.js\");\nfunction _slicedToArray(arr, i) { return _arrayWithHoles(arr) || _iterableToArrayLimit(arr, i) || _unsupportedIterableToArray(arr, i) || _nonIterableRest(); }\n\nfunction _nonIterableRest() { throw new TypeError(\"Invalid attempt to destructure non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); }\n\nfunction _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === \"string\") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === \"Object\" && o.constructor) n = o.constructor.name; if (n === \"Map\" || n === \"Set\") return Array.from(o); if (n === \"Arguments\" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }\n\nfunction _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }\n\nfunction _iterableToArrayLimit(arr, i) { var _i = arr == null ? null : typeof Symbol !== \"undefined\" && arr[Symbol.iterator] || arr[\"@@iterator\"]; if (_i == null) return; var _arr = []; var _n = true; var _d = false; var _s, _e; try { for (_i = _i.call(arr); !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i[\"return\"] != null) _i[\"return\"](); } finally { if (_d) throw _e; } } return _arr; }\n\nfunction _arrayWithHoles(arr) { if (Array.isArray(arr)) return arr; }\n\n\n\n\n\n\n\nvar SignUp = function SignUp() {\n  var _useState = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(\"\"),\n      _useState2 = _slicedToArray(_useState, 2),\n      username = _useState2[0],\n      setUsername = _useState2[1];\n\n  var _useState3 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(\"\"),\n      _useState4 = _slicedToArray(_useState3, 2),\n      nickname = _useState4[0],\n      setNickname = _useState4[1];\n\n  var _useState5 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(\"\"),\n      _useState6 = _slicedToArray(_useState5, 2),\n      email = _useState6[0],\n      setEmail = _useState6[1];\n\n  var _useState7 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(\"\"),\n      _useState8 = _slicedToArray(_useState7, 2),\n      pwd = _useState8[0],\n      setPwd = _useState8[1];\n\n  var _useState9 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(\"\"),\n      _useState10 = _slicedToArray(_useState9, 2),\n      pwdConfirm = _useState10[0],\n      setConfirm = _useState10[1];\n\n  var _useState11 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(false),\n      _useState12 = _slicedToArray(_useState11, 2),\n      submit = _useState12[0],\n      setSubmit = _useState12[1];\n\n  var _useState13 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(false),\n      _useState14 = _slicedToArray(_useState13, 2),\n      sent = _useState14[0],\n      setSent = _useState14[1];\n\n  var _useState15 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(),\n      _useState16 = _slicedToArray(_useState15, 2),\n      pwdMatch = _useState16[0],\n      setPwdMatch = _useState16[1];\n\n  var _useState17 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(),\n      _useState18 = _slicedToArray(_useState17, 2),\n      usernameChecked = _useState18[0],\n      setUsernameCheck = _useState18[1];\n\n  var _useState19 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(),\n      _useState20 = _slicedToArray(_useState19, 2),\n      nicknameChecked = _useState20[0],\n      setNicknameCheck = _useState20[1];\n\n  var _useState21 = (0,react__WEBPACK_IMPORTED_MODULE_1__.useState)(false),\n      _useState22 = _slicedToArray(_useState21, 2),\n      requestReview = _useState22[0],\n      setRequestReview = _useState22[1];\n\n  var onRegisterClick = function onRegisterClick() {\n    setSubmit(true);\n\n    if (validateInput()) {\n      axios__WEBPACK_IMPORTED_MODULE_0___default()({\n        method: \"post\",\n        url: \"/api/signup\",\n        headers: {\n          \"Content-Type\": \"application/json\"\n        },\n        data: {\n          username: username,\n          nickname: nickname,\n          email: email,\n          password1: pwd,\n          password2: pwdConfirm\n        }\n      })[\"catch\"](function (err) {\n        return err.response.status === 400 ? setRequestReview(\"비밀번호가 보안에 취약합니다. 다른 비밀번호를 입력해주세요.\") : err.response;\n      }).then(function (response) {\n        if (response.status < 400) {\n          setSent(true);\n        }\n      });\n    }\n  };\n\n  var validateInput = function validateInput() {\n    if (username === \"\" || email === \"\" || pwd === \"\" || pwdConfirm === \"\") {\n      setRequestReview(\"모든 영역은 필수 입력입니다.\");\n      return false;\n    } else if (!usernameChecked || !nicknameChecked) {\n      setRequestReview(\"아이디와 필명의 중복 확인이 필요합니다.\");\n      return false;\n    } else if (!pwdMatch) {\n      setRequestReview(\"비밀번호가 일치하지 않습니다.\");\n      return false;\n    } else if (pwd.length < 8) {\n      setRequestReview(\"비밀번호가 너무 짧습니다.\");\n      return false;\n    } else if (pwd.match(/^\\d+$/)) {\n      setRequestReview(\"비밀번호가 숫자로만 이루어져 있습니다.\");\n      return false;\n    }\n\n    return true;\n  };\n\n  var checkDuplicate = function checkDuplicate(type, target, setFunc) {\n    if (target === \"\") setFunc(false);else {\n      axios__WEBPACK_IMPORTED_MODULE_0___default()({\n        method: \"post\",\n        url: \"/api/\".concat(type, \"-check\"),\n        headers: {\n          \"Content-Type\": \"application/json\"\n        },\n        data: {\n          username: \"\".concat(target),\n          nickname: \"\".concat(target)\n        }\n      })[\"catch\"](function (error) {\n        if (error.response.status === 409) {\n          setFunc(false);\n          return error.response;\n        }\n      }).then(function (response) {\n        if (response.status < 400) {\n          setFunc(true);\n        }\n      });\n    }\n  };\n\n  (0,react__WEBPACK_IMPORTED_MODULE_1__.useEffect)(function () {\n    if (pwdConfirm === \"\") setPwdMatch();else if (pwd === pwdConfirm) setPwdMatch(true);else setPwdMatch(false);\n  }, [pwdConfirm]);\n  return sent ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_styles_globalStyles__WEBPACK_IMPORTED_MODULE_3__.Background, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.AfterSent, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.SentMessage, null, \"\\uC785\\uB825\\uD574\\uC8FC\\uC2E0 \\uC774\\uBA54\\uC77C\\uB85C \\uC778\\uC99D \\uB9C1\\uD06C\\uB97C \\uBCF4\\uB0B4\\uB4DC\\uB838\\uC2B5\\uB2C8\\uB2E4. \\uC778\\uC99D \\uD6C4 \\uB85C\\uADF8\\uC778 \\uAC00\\uB2A5\\uD569\\uB2C8\\uB2E4.\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.ToLoginLink, {\n    to: \"/login\"\n  }, \"\\uB85C\\uADF8\\uC778 \\uD558\\uAE30\"))) : /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_styles_globalStyles__WEBPACK_IMPORTED_MODULE_3__.Background, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterSection, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterForm, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_Components_csrftoken__WEBPACK_IMPORTED_MODULE_2__[\"default\"], null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterLegend, null, \"\\uD68C\\uC6D0\\uAC00\\uC785\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterFieldset, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputLabel, {\n    htmlFor: \"username\"\n  }, \"\\uC544\\uC774\\uB514\", \" \", usernameChecked === false ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.DuplicateMessage, null, \"\\uD574\\uB2F9 \\uC544\\uC774\\uB514\\uB294 \\uC0AC\\uC6A9\\uD560 \\uC218 \\uC5C6\\uC2B5\\uB2C8\\uB2E4.\") : \"\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterInput, {\n    type: \"text\",\n    name: \"username\",\n    placeholder: \"\\uACF5\\uBC31\\uC5C6\\uC774 \\uC601\\uBB38, \\uC22B\\uC790 \\uD3EC\\uD568 6-12\\uC790\",\n    duplicateChecked: usernameChecked,\n    onChange: function onChange(e) {\n      setUsername(e.target.value);\n      setUsernameCheck();\n      setRequestReview(\"\");\n    }\n  }), !usernameChecked ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.DuplicateCheckButton, {\n    onClick: function onClick() {\n      checkDuplicate(\"username\", username, setUsernameCheck);\n    }\n  }, \"\\uC911\\uBCF5 \\uD655\\uC778\") : /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.DuplicateConfirm, null, \"\\uC0AC\\uC6A9 \\uAC00\\uB2A5\\uD569\\uB2C8\\uB2E4\"))), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputLabel, {\n    htmlFor: \"nickname\"\n  }, \"\\uD544\\uBA85\", \" \", nicknameChecked === false ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.DuplicateMessage, null, \"\\uD574\\uB2F9 \\uD544\\uBA85\\uC740 \\uC0AC\\uC6A9\\uD560 \\uC218 \\uC5C6\\uC2B5\\uB2C8\\uB2E4.\") : \"\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterInput, {\n    type: \"text\",\n    name: \"nickname\",\n    placeholder: \"\\uAE00\\uC744 \\uC791\\uC131 \\uC2DC \\uD45C\\uC2DC\\uB418\\uB294 \\uC774\\uB984\",\n    duplicateChecked: nicknameChecked,\n    onChange: function onChange(e) {\n      setNickname(e.target.value);\n      setNicknameCheck();\n      setRequestReview(\"\");\n    }\n  }), !nicknameChecked ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.DuplicateCheckButton, {\n    onClick: function onClick() {\n      checkDuplicate(\"nickname\", nickname, setNicknameCheck);\n    }\n  }, \"\\uC911\\uBCF5 \\uD655\\uC778\") : /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.DuplicateConfirm, null, \"\\uC0AC\\uC6A9 \\uAC00\\uB2A5\\uD569\\uB2C8\\uB2E4\"))), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputLabel, {\n    htmlFor: \"username\"\n  }, \"\\uC774\\uBA54\\uC77C\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterInput, {\n    type: \"email\",\n    name: \"email\",\n    placeholder: \"ex. example@iieum.com\",\n    onChange: function onChange(e) {\n      setEmail(e.target.value);\n      setRequestReview(\"\");\n    }\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputLabel, {\n    htmlFor: \"password\"\n  }, \"\\uBE44\\uBC00\\uBC88\\uD638\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterInput, {\n    type: \"password\",\n    name: \"password\",\n    placeholder: \"\\uACF5\\uBC31\\uC5C6\\uC774 \\uC601\\uBB38, \\uC22B\\uC790 \\uD3EC\\uD568 8-20\\uC790\",\n    duplicateChecked: pwdMatch,\n    onChange: function onChange(e) {\n      setPwd(e.target.value);\n      setRequestReview(\"\");\n    }\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.InputLabel, {\n    htmlFor: \"passwordConfirm\"\n  }, \"\\uBE44\\uBC00\\uBC88\\uD638 \\uD655\\uC778\", pwdMatch === false ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.DuplicateMessage, null, \"\\uBE44\\uBC00\\uBC88\\uD638\\uAC00 \\uC77C\\uCE58\\uD558\\uC9C0 \\uC54A\\uC2B5\\uB2C8\\uB2E4.\") : \"\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterInput, {\n    type: \"password\",\n    name: \"passwordConfirm\",\n    placeholder: \"\\uBE44\\uBC00\\uBC88\\uD638 \\uC7AC\\uD655\\uC778\",\n    duplicateChecked: pwdMatch,\n    onChange: function onChange(e) {\n      setConfirm(e.target.value);\n      setRequestReview(\"\");\n    }\n  }))), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterBtnContainer, null, requestReview ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(react__WEBPACK_IMPORTED_MODULE_1__.Fragment, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.DuplicateMessage, null, requestReview), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(\"br\", null)) : \"\", submit ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.SubmitMessage, null, \"\\uCC98\\uB9AC \\uC911\\uC785\\uB2C8\\uB2E4...\") : /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(react__WEBPACK_IMPORTED_MODULE_1__.Fragment, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.RegisterBtn, {\n    onClick: function onClick() {\n      return onRegisterClick();\n    }\n  }, \"\\uAC00\\uC785\\uD558\\uAE30\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_1__.createElement(_SignUp_elements__WEBPACK_IMPORTED_MODULE_4__.ToLoginLink, {\n    to: \"/login\"\n  }, \"\\uB85C\\uADF8\\uC778\\uD558\\uAE30\"))))));\n};\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (SignUp);\n\n//# sourceURL=webpack://frontend/./src/Containers/SignUp/SignUp.js?");
 
 /***/ }),
 
@@ -983,7 +983,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"Navbar\": () => (/* reexport safe */ _Navbar_Navbar__WEBPACK_IMPORTED_MODULE_0__[\"default\"]),\n/* harmony export */   \"Home\": () => (/* reexport safe */ _Home_Home__WEBPACK_IMPORTED_MODULE_1__[\"default\"]),\n/* harmony export */   \"MyPage\": () => (/* reexport safe */ _MyPage_MyPage__WEBPACK_IMPORTED_MODULE_2__[\"default\"]),\n/* harmony export */   \"New\": () => (/* reexport safe */ _New_New__WEBPACK_IMPORTED_MODULE_3__[\"default\"]),\n/* harmony export */   \"SignUp\": () => (/* reexport safe */ _SignUp_SignUp__WEBPACK_IMPORTED_MODULE_4__[\"default\"]),\n/* harmony export */   \"Login\": () => (/* reexport safe */ _Login_Login__WEBPACK_IMPORTED_MODULE_5__[\"default\"]),\n/* harmony export */   \"EmailConfirmed\": () => (/* reexport safe */ _EmailConfirmed_EmailConfirmed__WEBPACK_IMPORTED_MODULE_6__[\"default\"]),\n/* harmony export */   \"Footer\": () => (/* reexport safe */ _Footer_Footer__WEBPACK_IMPORTED_MODULE_7__[\"default\"]),\n/* harmony export */   \"PostCardList\": () => (/* reexport safe */ _PostCardList_PostCardList__WEBPACK_IMPORTED_MODULE_8__[\"default\"]),\n/* harmony export */   \"Explore\": () => (/* reexport safe */ _Explore_Explore__WEBPACK_IMPORTED_MODULE_9__[\"default\"]),\n/* harmony export */   \"Search\": () => (/* reexport safe */ _Search_Search__WEBPACK_IMPORTED_MODULE_10__[\"default\"]),\n/* harmony export */   \"Change\": () => (/* reexport safe */ _Change_Change__WEBPACK_IMPORTED_MODULE_11__[\"default\"])\n/* harmony export */ });\n/* harmony import */ var _Navbar_Navbar__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./Navbar/Navbar */ \"./src/Containers/Navbar/Navbar.js\");\n/* harmony import */ var _Home_Home__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./Home/Home */ \"./src/Containers/Home/Home.js\");\n/* harmony import */ var _MyPage_MyPage__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./MyPage/MyPage */ \"./src/Containers/MyPage/MyPage.js\");\n/* harmony import */ var _New_New__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./New/New */ \"./src/Containers/New/New.js\");\n/* harmony import */ var _SignUp_SignUp__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./SignUp/SignUp */ \"./src/Containers/SignUp/SignUp.js\");\n/* harmony import */ var _Login_Login__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ./Login/Login */ \"./src/Containers/Login/Login.js\");\n/* harmony import */ var _EmailConfirmed_EmailConfirmed__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ./EmailConfirmed/EmailConfirmed */ \"./src/Containers/EmailConfirmed/EmailConfirmed.js\");\n/* harmony import */ var _Footer_Footer__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! ./Footer/Footer */ \"./src/Containers/Footer/Footer.js\");\n/* harmony import */ var _PostCardList_PostCardList__WEBPACK_IMPORTED_MODULE_8__ = __webpack_require__(/*! ./PostCardList/PostCardList */ \"./src/Containers/PostCardList/PostCardList.js\");\n/* harmony import */ var _Explore_Explore__WEBPACK_IMPORTED_MODULE_9__ = __webpack_require__(/*! ./Explore/Explore */ \"./src/Containers/Explore/Explore.js\");\n/* harmony import */ var _Search_Search__WEBPACK_IMPORTED_MODULE_10__ = __webpack_require__(/*! ./Search/Search */ \"./src/Containers/Search/Search.js\");\n/* harmony import */ var _Change_Change__WEBPACK_IMPORTED_MODULE_11__ = __webpack_require__(/*! ./Change/Change */ \"./src/Containers/Change/Change.js\");\n\n\n\n\n\n\n\n\n\n\n\n\n\n//# sourceURL=webpack://frontend/./src/Containers/index.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"Navbar\": () => (/* reexport safe */ _Navbar_Navbar__WEBPACK_IMPORTED_MODULE_0__[\"default\"]),\n/* harmony export */   \"Home\": () => (/* reexport safe */ _Home_Home__WEBPACK_IMPORTED_MODULE_1__[\"default\"]),\n/* harmony export */   \"MyPage\": () => (/* reexport safe */ _MyPage_MyPage__WEBPACK_IMPORTED_MODULE_2__[\"default\"]),\n/* harmony export */   \"New\": () => (/* reexport safe */ _New_New__WEBPACK_IMPORTED_MODULE_3__[\"default\"]),\n/* harmony export */   \"SignUp\": () => (/* reexport safe */ _SignUp_SignUp__WEBPACK_IMPORTED_MODULE_4__[\"default\"]),\n/* harmony export */   \"Login\": () => (/* reexport safe */ _Login_Login__WEBPACK_IMPORTED_MODULE_5__[\"default\"]),\n/* harmony export */   \"EmailConfirmed\": () => (/* reexport safe */ _EmailConfirmed_EmailConfirmed__WEBPACK_IMPORTED_MODULE_6__[\"default\"]),\n/* harmony export */   \"Footer\": () => (/* reexport safe */ _Footer_Footer__WEBPACK_IMPORTED_MODULE_7__[\"default\"]),\n/* harmony export */   \"PostCardList\": () => (/* reexport safe */ _PostCardList_PostCardList__WEBPACK_IMPORTED_MODULE_8__[\"default\"]),\n/* harmony export */   \"Explore\": () => (/* reexport safe */ _Explore_Explore__WEBPACK_IMPORTED_MODULE_9__[\"default\"]),\n/* harmony export */   \"Search\": () => (/* reexport safe */ _Search_Search__WEBPACK_IMPORTED_MODULE_10__[\"default\"]),\n/* harmony export */   \"Change\": () => (/* reexport safe */ _Change_Change__WEBPACK_IMPORTED_MODULE_11__[\"default\"]),\n/* harmony export */   \"MusicCardGrid\": () => (/* reexport safe */ _MusicCardGrid_MusicCardGrid__WEBPACK_IMPORTED_MODULE_12__[\"default\"]),\n/* harmony export */   \"QuestionCardGrid\": () => (/* reexport safe */ _QuestionCardGrid_QuestionCardGrid__WEBPACK_IMPORTED_MODULE_13__[\"default\"])\n/* harmony export */ });\n/* harmony import */ var _Navbar_Navbar__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./Navbar/Navbar */ \"./src/Containers/Navbar/Navbar.js\");\n/* harmony import */ var _Home_Home__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./Home/Home */ \"./src/Containers/Home/Home.js\");\n/* harmony import */ var _MyPage_MyPage__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./MyPage/MyPage */ \"./src/Containers/MyPage/MyPage.js\");\n/* harmony import */ var _New_New__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./New/New */ \"./src/Containers/New/New.js\");\n/* harmony import */ var _SignUp_SignUp__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./SignUp/SignUp */ \"./src/Containers/SignUp/SignUp.js\");\n/* harmony import */ var _Login_Login__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ./Login/Login */ \"./src/Containers/Login/Login.js\");\n/* harmony import */ var _EmailConfirmed_EmailConfirmed__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ./EmailConfirmed/EmailConfirmed */ \"./src/Containers/EmailConfirmed/EmailConfirmed.js\");\n/* harmony import */ var _Footer_Footer__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! ./Footer/Footer */ \"./src/Containers/Footer/Footer.js\");\n/* harmony import */ var _PostCardList_PostCardList__WEBPACK_IMPORTED_MODULE_8__ = __webpack_require__(/*! ./PostCardList/PostCardList */ \"./src/Containers/PostCardList/PostCardList.js\");\n/* harmony import */ var _Explore_Explore__WEBPACK_IMPORTED_MODULE_9__ = __webpack_require__(/*! ./Explore/Explore */ \"./src/Containers/Explore/Explore.js\");\n/* harmony import */ var _Search_Search__WEBPACK_IMPORTED_MODULE_10__ = __webpack_require__(/*! ./Search/Search */ \"./src/Containers/Search/Search.js\");\n/* harmony import */ var _Change_Change__WEBPACK_IMPORTED_MODULE_11__ = __webpack_require__(/*! ./Change/Change */ \"./src/Containers/Change/Change.js\");\n/* harmony import */ var _MusicCardGrid_MusicCardGrid__WEBPACK_IMPORTED_MODULE_12__ = __webpack_require__(/*! ./MusicCardGrid/MusicCardGrid */ \"./src/Containers/MusicCardGrid/MusicCardGrid.js\");\n/* harmony import */ var _QuestionCardGrid_QuestionCardGrid__WEBPACK_IMPORTED_MODULE_13__ = __webpack_require__(/*! ./QuestionCardGrid/QuestionCardGrid */ \"./src/Containers/QuestionCardGrid/QuestionCardGrid.js\");\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n//# sourceURL=webpack://frontend/./src/Containers/index.js?");
 
 /***/ }),
